### 2023-04-26 #5412 @Kevin-Mattheus-Moerman
@blaiszik I am the AEiC for this journal and here to help process initial steps. Please work on the below before we proceed:

- [ ] Please fix the paper compilation. The above error looks like it relates to invalid (all zeros) ORCID entries. All ORCIDs should be valid. For those that do not have an ORCID please remove the entire ORCID line.
- [ ] Please check the above potentially missing DOIs. You can amend your .bib file to correct these.

After you've fixed these you can call `@editorialbot check references` to check the DOIs, and use `@editorialbot generate pdf` to recompile the draft paper.
