### 2023-04-26 #5412 @Kevin-Mattheus-Moerman
@blaiszik given the relatively small size (in terms of lines of code) of this submission, I have just triggered a scope review by the editorial board, to see if this work conforms to out [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This review should take about 2 weeks to complete.
