+++ foundry/foundry.py
@@ -407,7 +407,8 @@ class Foundry(FoundryMetadata):
                 endpoint_auth_clients={endpoint_id: AuthClient(authorizer=self.auths[scope])}
             )
             # upload (ie publish) data to endpoint
-            globus_data_source, rule_id = upload_to_endpoint(pub_auths, https_data_path, endpoint_id)
+            # globus_data_source, rule_id = upload_to_endpoint(pub_auths, https_data_path, endpoint_id)
+            globus_data_source = upload_to_endpoint(pub_auths, https_data_path, endpoint_id)
         # set Globus data source URL with MDF
         self.connect_client.add_data_source(globus_data_source)
         # set dataset name using the title if an abbreviated short_name isn't specified
@@ -420,9 +421,9 @@ class Foundry(FoundryMetadata):
         else:
             res = None
 
-        # if uploaded by HTTPS, delete ACL rule after dataset submission is complete
-        if https_data_path and rule_id:
-            self.transfer_client.delete_endpoint_acl_rule(endpoint_id, rule_id)
+        # # if uploaded by HTTPS, delete ACL rule after dataset submission is complete
+        # if https_data_path and rule_id:
+        #     self.transfer_client.delete_endpoint_acl_rule(endpoint_id, rule_id)
 
         return res
 
