### 2023-05-03 ff82896
add missing scopes
