+++ foundry/foundry.py
@@ -80,11 +80,13 @@ class Foundry(FoundryMetadata):
                     "petrel",
                     "transfer",
                     "dlhub",
-                    "funcx",
                     "openid",
+                    "email",
+                    "profile",
                     "https://auth.globus.org/scopes/facd7ccc-c5f4-42aa-916b-a0e270e2c2a9/all",  # funcx
                     "https://auth.globus.org/scopes/f10a69a9-338c-4e5b-baa1-0dc92359ab47/https",  # Eagle HTTPS
                     "https://auth.globus.org/scopes/82f1b5c6-6e9b-11e5-ba47-22000b92c6ec/https",  # NCSA HTTPS
+                    "https://auth.globus.org/scopes/d31d4f5d-be37-4adc-a761-2f716b7af105/action_all", # Globus Search Lambda
                 ]
             self.auths = mdf_toolbox.login(
                 services=services,
