### 2023-05-03 11c6459
Add search lambda authorizer (sl_authorizer) to dlhub_client instantiation
