### 2023-05-03 b2c7598
removed unnecessary scopes
