### 2023-05-04 9c55200
Update publishing notebook and minor bugfixes (#336)
* update publishing notebook example to use HTTPS upload primarily, along with minor fixes

* add https upload methods and data

* fix function call to publish_dataset

* remove ACL rule code to fix error issue

* update globus images in notebook

* remove commented code

* add missing scopes

* appease flake overlords

* Add search lambda authorizer (sl_authorizer) to dlhub_client
instantiation

* removed unnecessary scopes

* update curation info in notebook

---------

Co-authored-by: Ben Blaiszik <blaiszik@uchicago.edu>
Co-authored-by: i
[…]
