### 2023-05-04 fdc6d50
update to 0.6.0 for HTTPS pub
