### 2023-05-08 f2f9a58
GITBOOK-104: add HTTPS as default; fix data publishing example; update publish_dataset method
