### 2023-05-09 d35e64c
GITBOOK-105: fix pub example to point to folder
