### 2023-05-10 #5412 @Kevin-Mattheus-Moerman
@fabian-s although this work is applied to material science, I thought you might be a good fit given your machine learning background. Let me know if you can help edit. Thanks!
