### 2023-05-10 #5412 @fabian-s
Sorry, I have to decline this time -- don't know enough python, application area is far from my expertise and I am under a lot of pressure at my day job ATM.
