### 2023-05-11 6a0e132
Upload Foundry class load() function to default download using https
