### 2023-05-11 3673fdf
Update setup.py
Fix version number for pyPI deploy
