### 2023-05-16 #5467 @editorialbot
[REVIEW]: Foundry-ML - Software and Services to Simplify Access to Machine Learning Datasets in Materials Science
**Submitting author:** @blaiszik (<a href="http://orcid.org/0000-0002-5326-4902">Ben Blaiszik</a>)
**Repository:** https://github.com/MLMI2-CSSI/foundry
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.7.4
