### 2023-05-25 3c84e47
adding ability to specify splits for loading
