### 2023-05-26 0d01a7e
incorporate load() to foundry.__init__()
