### 2023-05-30 f85b665
Split specification (#344)
* adding ability to specify splits for loading

* refining test

* Update splits_to_load --> splits

---------

Co-authored-by: blaiszik <blaiszik@uchicago.edu>
