### 2023-05-30 9572b6c
automating api documentation using github action (#342)
