+++ docs/foundry.foundry.md
@@ -1,0 +1,415 @@
+<!-- markdownlint-disable -->
+
+<a href="https://github.com/MLMI2-CSSI/foundry/tree/main/foundry/foundry.py#L0"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+# <kbd>module</kbd> `foundry.foundry`
+
+
+
+
+
+
+---
+
+<a href="https://github.com/MLMI2-CSSI/foundry/tree/main/foundry/foundry.py#L34"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+## <kbd>class</kbd> `Foundry`
+Foundry Client Base Class 
+
+**TODO:**
+ 
+------- Add Docstring 
+
+<a href="https://github.com/MLMI2-CSSI/foundry/tree/main/foundry/foundry.py#L52"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+### <kbd>method</kbd> `__init__`
+
+```python
+__init__(
+    no_browser=False,
+    no_local_server=False,
+    index='mdf',
+    authorizers=None,
+    **data
+)
+```
+
+Initialize a Foundry client 
+
+**Args:**
+ 
+ - <b>`no_browser`</b> (bool):   Whether to open the browser for the Globus Auth URL. 
+ - <b>`no_local_server`</b> (bool):  Whether a local server is available.  This should be `False` when on remote server (e.g., Google Colab ). 
+ - <b>`index`</b> (str):  Index to use for search and data publication. Choices `mdf` or `mdf-test` 
+ - <b>`authorizers`</b> (dict):  A dictionary of authorizers to use, following the `mdf_toolbox` format 
+ - <b>`data`</b> (dict):  Other arguments, e.g., results from an MDF search result that are used  to populate Foundry metadata fields 
+
+Returns 
+------- an initialized and authenticated Foundry client 
+
+
+
+
+---
+
+<a href="https://github.com/MLMI2-CSSI/foundry/tree/main/foundry/foundry.py#L452"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+### <kbd>method</kbd> `check_status`
+
+```python
+check_status(source_id, short=False, raw=False)
+```
+
+Check the status of your submission. 
+
+
+
+**Arguments:**
+ 
+ - <b>`source_id`</b> (str):  The ``source_id`` (``source_name`` + version information) of the  submission to check. Returned in the ``res`` result from ``publish()`` via MDF Connect Client. 
+ - <b>`short`</b> (bool):  When ``False``, will print a status summary containing  all of the status steps for the dataset.  When ``True``, will print a short finished/processing message,  useful for checking many datasets' status at once. 
+ - <b>`**Default`</b>: ** ``False`` 
+ - <b>`raw`</b> (bool):  When ``False``, will print a nicely-formatted status summary.  When ``True``, will return the full status result.  For direct human consumption, ``False`` is recommended. 
+ - <b>`**Default`</b>: ** ``False`` 
+
+
+
+**Returns:**
+ 
+ - <b>`If ``raw`` is ``True``, *dict*`</b>:  The full status result. 
+
+---
+
+<a href="https://github.com/MLMI2-CSSI/foundry/tree/main/foundry/foundry.py#L481"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+### <kbd>method</kbd> `configure`
+
+```python
+configure(**kwargs)
+```
+
+Set Foundry config Keyword Args:  file (str): Path to the file containing  (default: self.config.metadata_file) 
+
[…]
