+++ docs/foundry.utils.md
@@ -1,0 +1,45 @@
+<!-- markdownlint-disable -->
+
+<a href="https://github.com/MLMI2-CSSI/foundry/tree/main/foundry/utils.py#L0"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+# <kbd>module</kbd> `foundry.utils`
+
+
+
+
+
+---
+
+<a href="https://github.com/MLMI2-CSSI/foundry/tree/main/foundry/utils.py#L4"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+## <kbd>function</kbd> `is_pandas_pytable`
+
+```python
+is_pandas_pytable(group)
+```
+
+
+
+
+
+
+---
+
+<a href="https://github.com/MLMI2-CSSI/foundry/tree/main/foundry/utils.py#L11"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+## <kbd>function</kbd> `is_doi`
+
+```python
+is_doi(string: str)
+```
+
+
+
+
+
+
+
+
+---
+
+_This file was automatically generated via [lazydocs](https://github.com/ml-tooling/lazydocs)._
