+++ docs/foundry.loaders.md
@@ -1,0 +1,16 @@
+<!-- markdownlint-disable -->
+
+<a href="https://github.com/MLMI2-CSSI/foundry/tree/main/foundry/loaders/__init__.py"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+# <kbd>module</kbd> `foundry.loaders`
+
+
+
+
+
+
+
+
+---
+
+_This file was automatically generated via [lazydocs](https://github.com/ml-tooling/lazydocs)._
