### 2023-05-31 efc7f68
Update README.md with contributing instructions (#357)
* Update README.md with contributing instructions

* Update PR language
