### 2023-06-03 #5467 @blaiszik
Thank you @marshallmcdonnell! There are a series of examples in the `examples` folder that should run easily on Google Colab. If you run locally and don't want to install the Globus Connect personal endpoint just set the variable `globus=False`. We will go  back and make this the default ASAP since it gets people up and running without additional installs or reliance on Colab.

https://github.com/MLMI2-CSSI/foundry/tree/main/examples
