### 2023-06-08 68f3287
Load on init (#358)
* incorporate load() to foundry.__init__()

* merging in split specification

* flake fixes

* removed blank line
