### 2023-06-15 MLMI2-CSSI/foundry#372 @marshallmcdonnell
"Quick start" in README won't work w/o Globus Connect Personal
NOTE: This is related to JOSS review https://github.com/openjournals/joss-reviews/issues/5467

In the "Quick Start", the `globus=True` is set for the `Foundry.load()` method.

Yet, this requires a Globus Connect Personal (or maybe Globus Connect Server) endpoint to be setup already.

There is no documentation link to set that up in the README.

Per the [comment](https://github.com/openjournals/joss-reviews/issues/5467#issuecomment-1575192043) from @blaiszik, can easily disable this using `globus=False`.

Probably want the "quickstart" to work without a Globus endpoint (and `globus=False` by default)

**Describe the solution you'd like**
Just use the default so a User doesn't have to configure a Globus endpoint for an introduction the package.

Keep `globus=True` but add a "note" that you can set `globus=False` if you would like to not use a Globus endpoint

**Additional context**
N/A
