### 2023-06-15 ec7ba1c
Adds note for quickstart globus set to false
