### 2023-06-15 #5467 @marshallmcdonnell
Quick note:

For contributions:
> Contribution and authorship: Has the submitting author (@blaiszik) made major contributions to the software? *Does the full list of paper authors seem appropriate and complete?*

I wanted to check on something.
I went to the contributors list for Foundry [here](https://github.com/MLMI2-CSSI/foundry/graphs/contributors).

I checked this against the author list.
There are contributors on the GitHub list not added as authors on this list:
  - Steve Wangen: https://github.com/blue442
  - Braeden Cullen: https://github.com/BraedenCu
  - Nathaniel Martinez: https://github.com/nmartinez233
  - Eric Blau (sp?): https://github.com/ericblau
  - Sterling G. Baird: https://github.com/sgbaird
  - Cora Schneck: https://github.com/cyschneck
  - Nathan Pruyne: https://github.com/NathanPruyne
  - Ryan Chard: https://github.com/ryanchard

  There are also authors included in the author list that are not contributors. Yet, I'm not so worried about that.

My main concern is should the contributors to the source code listed above not be added to the author list?
