### 2023-06-15 MLMI2-CSSI/foundry#375 @marshallmcdonnell
Broken links in documentation
NOTE: This is related to JOSS review https://github.com/openjournals/joss-reviews/issues/5467

**Describe the bug**
Broken links in docs throughout code base:

1. Following link is broken in the docs for publishing datasets:
https://github.com/MLMI2-CSSI/foundry/blob/main/docs/publishing/publishing-datasets.md?plain=1#L11

This line has the following link that does not exist:
```
[Skip to the publication example notebook.](https://github.com/MLMI2-CSSI/foundry/blob/master/examples/foundry_publication_example.ipynb)
```

2. Broken link for contributing guide:
https://github.com/MLMI2-CSSI/foundry/blob/main/docs/how-to-contribute/contributing.md?plain=1#L34

This line has the following link that does not exist:
```
* Tests should follow [testing best practices](https://www..org/community/contribute/tests)
```

**To Reproduce**
Try to navigate to these links.

**Expected behavior**
For (1), I expect that it is now a new name for the notebook and probably found here now?
https://github.com/MLMI2-CSSI/foundry/tree/main/examples/publishing-guides
