### 2023-06-15 MLMI2-CSSI/foundry#376 @marshallmcdonnell
Mixed docstring styles in code base
NOTE: This is related to JOSS review https://github.com/openjournals/joss-reviews/issues/5467

Currently, I am coming across multiple docstring styles in the code base.

Is this intentional?

Example:
  - [Google docstring style](https://github.com/google/styleguide/blob/gh-pages/pyguide.md#38-comments-and-docstrings): https://github.com/MLMI2-CSSI/foundry/blob/main/foundry/utils.py#L21-L24
  - (_sort of_) numpydoc docstring style(https://numpydoc.readthedocs.io/en/v1.1.0/format.html): https://github.com/MLMI2-CSSI/foundry/blob/main/foundry/https_upload.py#L31-L34

**Describe the solution you'd like**
Maybe add a docstring format used by Foundry in the contribution guide?
  - https://github.com/MLMI2-CSSI/foundry/blob/main/docs/how-to-contribute/contributing.md

Consistent docstring format in code

NOTE: Related to the docstrings, the following `upload_to_endpoint` function and the docstring args do not match (https://github.com/MLMI2-CSSI/foundry/blob/main/foundry/https_upload.py#L19-L29). Missing `dest_parent` and `dest_child` in the docstrings.
