### 2023-06-15 #5467 @marshallmcdonnell
Quick note:

> Functionality: Have the functional claims of the software been confirmed?
I'm hesitant to check this off since I was having trouble running the notebooks in Google Collab (mainly the `DefectTrack`).

I let it spin for 30 minutes on the `result = f.run("jwei74_wisc/DefectTrack_in_situ_TEM_videos", {'images':imgs_original})` cell but finally gave up.

I can assume that these notebooks work since the output is saved and stored in the current `main` branch source code files.

Since the authors heavily rely on notebooks for the examples, have you considered using a tool like `nbdev` to run actual tests on these notebooks to ensure the software examples can be sustained with continuing development?
  - https://github.com/fastai/nbdev

  Disclaimer: I haven't used `nbdev` extensively so only suggesting this if it helps.
