### 2023-06-15 #5467 @marshallmcdonnell
For this:

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I see that GitHub Actions is used for CI testing. I also see that there are GitHub secrets used for what looks like Globus credentials to run the tests.

I did try to run the tests locally but have the following errors:
```
FAILED tests/test_foundry.py::test_download_globus - globus_sdk.exc.base.GlobusSDKUsageError: destination_endpoint is required
FAILED tests/test_foundry.py::test_globus_dataframe_load - globus_sdk.exc.base.GlobusSDKUsageError: destination_endpoint is required
FAILED tests/test_foundry.py::test_publish_with_https - assert False
FAILED tests/test_foundry.py::test_publish_with_globus - assert False
```

Yet it appears all of these tests are decorated with:
```
@pytest.mark.skipif(bool(is_gha), reason="Not run as part of GHA CI")
```

This is fine. But there is no documentation in the `tests/README` to explain how one would run these tests or that they will fail locally without a globus endpoint enabled (I assume).

Finally, since these tests require a third-party application (Globus), perhaps these should be move to "integration tests"? And leave the current tests as proper "unit tests"? That way they can be run separately.

Just a suggestion for new developers.
