### 2023-06-15 #5467 @marshallmcdonnell
For this:

> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

I am using the proofs from https://github.com/openjournals/joss-reviews/issues/5467#issuecomment-1548909618 and these show that the citations are "broken" (i.e. `(yearofopendata?)` in the first paragraph of "Summary" on page 1, ` (Li67 et al., 2021; wei2021benchmark?), MIT (Schwalbe-Koda et al., 2021) ??,` in "Usage" section on page 3
