### 2023-06-15 #5467 @marshallmcdonnell
Overall, Foundry-ML seems like a great contribution to JOSS. I have left my review comments above.

For the source code, I've opened the following Issues:

- [ ] https://github.com/MLMI2-CSSI/foundry/issues/372
- [ ] https://github.com/MLMI2-CSSI/foundry/issues/373
- [ ] https://github.com/MLMI2-CSSI/foundry/issues/375
- [ ] https://github.com/MLMI2-CSSI/foundry/issues/376

And opened a small PR:
- https://github.com/MLMI2-CSSI/foundry/pull/374

I hope these are helpful!
