### 2023-06-26 ebe93ea
Validating metadata before publishing
