### 2023-06-26 5564660
remove arguments from Foundry object that are duplicated with base class
