### 2023-06-30 9f5a2e9
extract load from __init__ and restructure
