### 2023-07-06 1c360fb
refactoring for clarity
