+++ foundry/foundry.py
@@ -24,7 +24,6 @@ from .utils import _read_csv, _read_json, _read_excel
 from foundry.models import (
     FoundryMetadata,
     FoundryConfig,
-    FoundryDataset,
     FoundryBase
 )
 from foundry.https_download import download_file, recursive_ls
@@ -154,17 +153,23 @@ class Foundry(FoundryBase):
             force_login=False,
         )
 
-    def fetch_data(self, name, download=True, globus=True, verbose=False, authorizers=None, interval=10):
-        """Load the metadata for a Foundry dataset into the client
+    def fetch_data(self, name, metadata_only=False, globus=True, verbose=False, authorizers=None, interval=10):
+        """Load the data for a Foundry dataset into the client
         Args:
             name (str): Name of the foundry dataset
-            download (bool): If True, download the data associated with the package (default is True)
+            metadata_only (bool): If True, download only the metadata (not the full dataset)
+                associated with the package (default is False)
             globus (bool): If True, download using Globus, otherwise https
             verbose (bool): If True print additional debug information
             interval (int): How often to poll Globus to check if transfers are complete
 
         Returns:
             self
+
+        Todo:
+            * Add check for package existence during download of dataset
+            * Creating a new Foundry instance is a problematic way to update the metadata,
+                we should find a way to abstract this.
         """
 
         # handle empty dataset name (was returning all the datasets)
@@ -198,15 +203,12 @@ class Foundry(FoundryBase):
 
         del res["projects"][self.config.metadata_key]
 
-        # TODO: Creating a new Foundry instance is a problematic way to update the metadata,
-        # we should find a way to abstract this.
-
         self.dc = res['dc']
         self.mdf = res['mdf']
-        self.dataset = FoundryDataset(**res['dataset'])
+        self.metadata = FoundryMetadata(**res['dataset'])
 
-        if download:  # Add check for package existence
-            self.download(
+        if not metadata_only:
+            self.download_dataset(
                 interval=interval, globus=globus, verbose=verbose
             )
 
@@ -290,13 +292,13 @@ class Foundry(FoundryBase):
 
         # Handle splits if they exist. Return as a labeled dictionary of tuples
         try:
-            if self.dataset.splits:
+            if self.metadata.splits:
                 if not splits:
-                    for split in self.dataset.splits:
+                    for split in self.metadata.splits:
                         data[split.label] = self._load_data(file=split.path, source_id=source_id, globus=globus,
                                                             as_hdf5=as_hdf5)
                 else:
-                    for split in self.dataset.splits:
+                    for split in self.metadata.splits:
                         if split.label in splits:
                             splits.remove(split.label)
                             data[split.label] = self._load_data(file=split.path, source_id=source_id, globus=globus,
@@ -325,7 +327,7 @@ class Foundry(FoundryBase):
 
             buf = f'<h2>{title}</h2>{authors}<p>{DOI}</p>'
 
-            buf = f'{buf}<h3>Dataset</h3>{convert(json.loads(self.dataset.json(exclude={"dataframe"})))}'
+            buf = f'{buf}<h3>Dataset</h3>{convert(json.loads(self.metadata.json(exclude={"dataframe"})))}'
         return buf
 
     def get_citation(self) -> str:
@@ -509,7 +511,7 @@ class Foundry(FoundryBase):
         self.config = FoundryConfig(**kwargs)
         return self
 
-    def download(self, globus: bool = True, interval: int = 20, parallel_https: int = 4, verbose: bool = False) -> 'Foundry':
+    def download_dataset(self, globus: bool = True, interval: int = 20, parallel_https: int = 4, verbose: bool = False) -> 'Foundry':
         """Download a Foundry dataset
 
         Args:
@@ -526,10 +528,10 @@ class Foundry(FoundryBase):
[…]
