### 2023-07-21 764788f
remove globus from Foundry object instantiation
