### 2023-07-21 ddd995b
update call to fetch_data()
