### 2023-07-28 074857d
updating to allow instantiation with dataset name
