### 2023-07-28 6966030
adding docstrings, moving check for empty name
