### 2023-08-01 ccc2384
add functionality and testing around addition of links when publishing a dataset
