### 2023-08-01 8941a73
change so all links are added to MDF client at once
