+++ tests/test_foundry.py
@@ -207,33 +207,33 @@ def _delete_test_data(foundry_obj):
 
 
 def test_foundry_init():
-    f = Foundry(test_dataset, download=False, authorizers=auths)
+    f = Foundry(authorizers=auths)
     assert isinstance(f.forge_client, Forge)
     assert isinstance(f.connect_client, MDFConnectClient)
 
     if not is_gha:
         assert isinstance(f.dlhub_client, DLHubClient)
 
-        f2 = Foundry(test_dataset, download=False, authorizers=auths, no_browser=False, no_local_server=True)
+        f2 = Foundry(authorizers=auths, no_browser=False, no_local_server=True)
         assert isinstance(f2.dlhub_client, DLHubClient)
         assert isinstance(f2.forge_client, Forge)
         assert isinstance(f2.connect_client, MDFConnectClient)
 
-        f3 = Foundry(test_dataset, download=False, authorizers=auths, no_browser=True, no_local_server=False)
+        f3 = Foundry(authorizers=auths, no_browser=True, no_local_server=False)
         assert isinstance(f3.dlhub_client, DLHubClient)
         assert isinstance(f3.forge_client, Forge)
         assert isinstance(f3.connect_client, MDFConnectClient)
 
 
 def test_list():
-    f = Foundry(test_dataset, download=False, authorizers=auths)
+    f = Foundry(authorizers=auths)
     ds = f.list()
     assert isinstance(ds, pd.DataFrame)
     assert len(ds) > 0
 
 
 def test_search():
-    f = Foundry(test_dataset, download=False, authorizers=auths)
+    f = Foundry(authorizers=auths)
     q = "Elwood"
     ds = f.search(q)
 
@@ -245,20 +245,25 @@ def test_search():
 
 
 def test_metadata_pull():
-    f = Foundry(test_dataset, download=False, authorizers=auths)
+    f = Foundry(authorizers=auths)
+    assert f.dc == {}
+    f = f.load(test_dataset, download=False, authorizers=auths)
     assert f.dc["titles"][0]["title"] == expected_title
 
 
 def test_download_https():
-    f = Foundry(test_dataset, download=True, globus=False, authorizers=auths)
+    f = Foundry(authorizers=auths)
     _delete_test_data(f)
+    f = f.load(test_dataset, download=True, globus=False, authorizers=auths)
 
     assert f.dc["titles"][0]["title"] == expected_title
     _delete_test_data(f)
 
 
 def test_dataframe_load():
-    f = Foundry(test_dataset, download=True, globus=False, authorizers=auths)
+    f = Foundry(authorizers=auths)
+    _delete_test_data(f)
+    f = f.load(test_dataset, download=True, globus=False, authorizers=auths)
 
     res = f.load_data()
     X, y = res['train']
@@ -271,7 +276,9 @@ def test_dataframe_load():
 
 
 def test_dataframe_load_split():
-    f = Foundry(test_dataset, download=True, globus=False, authorizers=auths)
+    f = Foundry(authorizers=auths)
+    _delete_test_data(f)
+    f = f.load(test_dataset, download=True, globus=False, authorizers=auths)
 
     res = f.load_data(splits=['train'])
     X, y = res['train']
@@ -284,7 +291,9 @@ def test_dataframe_load_split():
 
 
 def test_dataframe_load_split_wrong_split_name():
-    f = Foundry(test_dataset, download=True, globus=False, authorizers=auths)
+    f = Foundry(authorizers=auths)
+    _delete_test_data(f)
+    f = f.load(test_dataset, download=True, globus=False, authorizers=auths)
 
     with pytest.raises(Exception) as exc_info:
[…]
