### 2023-11-03 b4d46a1
refactor foundry to separate foundry instance from dataset objects
