### 2023-11-03 b34f85f
fine tuning search functionality
