+++ foundry/foundry.py
@@ -1,16 +1,9 @@
 import h5py
-import json
 import mdf_toolbox
-from json2table import convert
-import numpy as np
 import pandas as pd
-from pydantic import ValidationError
 from typing import Any, Dict, List
 import logging
-import warnings
 import os
-from concurrent.futures import ThreadPoolExecutor, as_completed
-from tqdm.auto import tqdm
 
 from mdf_connect_client import MDFConnectClient
 from mdf_forge import Forge
@@ -22,14 +15,16 @@ from .utils import is_pandas_pytable, is_doi
 from .utils import _read_csv, _read_json, _read_excel
 
 from foundry.models import (
-    FoundryMetadata,
-    FoundryConfig,
+    FoundrySchema,
     FoundryDataset,
     FoundryBase
 )
-from foundry.https_download import download_file, recursive_ls
+
+from foundry.foundry_cache import FoundryCache
+
 from foundry.https_upload import upload_to_endpoint
 
+logging.basicConfig(format='%(levelname)s: %(message)s')
 logger = logging.getLogger(__name__)
 
 
@@ -50,23 +45,19 @@ class Foundry(FoundryBase):
     auths: Any
 
     def __init__(
-            self, name=None, no_browser=False, no_local_server=False, index="mdf", authorizers=None,
-            download=True, globus=True, verbose=False, metadata=None, interval=10,
+            self, no_browser=False, no_local_server=False, index="mdf", authorizers=None,
+            globus=True, verbose=False, interval=10,
             **data
     ):
         """Initialize a Foundry client
         Args:
-            name (str): Name of the foundry dataset. If not supplied, metadata will not be loaded into
-                    the Foundry object
             no_browser (bool):  Whether to open the browser for the Globus Auth URL.
             no_local_server (bool): Whether a local server is available.
                     This should be `False` when on remote server (e.g., Google Colab ).
             index (str): Index to use for search and data publication. Choices `mdf` or `mdf-test`
             authorizers (dict): A dictionary of authorizers to use, following the `mdf_toolbox` format
-            download (bool): If True, download the data associated with the package (default is True)
             globus (bool): If True, download using Globus, otherwise https
             verbose (bool): If True print additional debug information
-            metadata (dict): **For debug purposes.** A search result analog to prepopulate metadata.
             interval (int): How often to poll Globus to check if transfers are complete
             data (dict): Other arguments, e.g., results from an MDF search result that are used
                     to populate Foundry metadata fields
@@ -77,13 +68,7 @@ class Foundry(FoundryBase):
         super().__init__(**data)
         self.index = index
         self.auths = None
-
-        self.config = FoundryConfig(
-            dataframe_file="foundry_dataframe.json",
-            metadata_file="foundry_metadata.json",
-            local=False,
-            local_cache_dir="./data",
-        )
+        self.cache = FoundryCache()
 
         if authorizers:
             self.auths = authorizers
@@ -155,208 +140,126 @@ class Foundry(FoundryBase):
             force_login=False,
         )
 
-        if name is not None:
-            self._load(name=name,
-                       download=download,
-                       globus=globus,
-                       verbose=verbose,
-                       metadata=metadata,
-                       authorizers=authorizers,
-                       interval=interval)
+    def search(self, query: str = None, limit: int = None) -> [FoundryDataset]:
+        """Search available Foundry datasets
 
-    def _load(self, name, download=True, globus=True, verbose=False, metadata=None, authorizers=None, interval=None):
-        """Load the metadata for a Foundry dataset into the client
         Args:
-            name (str): Name of the foundry dataset
-            download (bool): If True, download the data associated with the package (default is True)
-            globus (bool): If True, download using Globus, otherwise https
[…]
