+++ tests/test_foundry.py
@@ -11,7 +11,7 @@ import requests
 import mdf_toolbox
 import pandas as pd
 from mdf_forge import Forge
-from foundry import Foundry
+from foundry import foundry
 from foundry.auth import PubAuths
 from foundry.https_upload import upload_to_endpoint
 from dlhub_sdk import DLHubClient
@@ -207,48 +207,74 @@ def _delete_test_data(foundry_obj):
 
 
 def test_foundry_init():
-    f = Foundry(test_dataset, download=False, authorizers=auths)
+    f = foundry.Foundry(authorizers=auths)
     assert isinstance(f.forge_client, Forge)
     assert isinstance(f.connect_client, MDFConnectClient)
 
     if not is_gha:
         assert isinstance(f.dlhub_client, DLHubClient)
 
-        f2 = Foundry(test_dataset, download=False, authorizers=auths, no_browser=False, no_local_server=True)
+        f2 = foundry.Foundry(download=False, authorizers=auths, no_browser=False, no_local_server=True)
         assert isinstance(f2.dlhub_client, DLHubClient)
         assert isinstance(f2.forge_client, Forge)
         assert isinstance(f2.connect_client, MDFConnectClient)
 
-        f3 = Foundry(test_dataset, download=False, authorizers=auths, no_browser=True, no_local_server=False)
+        f3 = foundry.Foundry(download=False, authorizers=auths, no_browser=True, no_local_server=False)
         assert isinstance(f3.dlhub_client, DLHubClient)
         assert isinstance(f3.forge_client, Forge)
         assert isinstance(f3.connect_client, MDFConnectClient)
 
 
 def test_list():
-    f = Foundry(test_dataset, download=False, authorizers=auths)
+    f = foundry.Foundry(authorizers=auths)
     ds = f.list()
-    assert isinstance(ds, pd.DataFrame)
+    # assert isinstance(ds, pd.DataFrame)
     assert len(ds) > 0
 
 
 def test_search():
-    f = Foundry(test_dataset, download=False, authorizers=auths)
+    f = foundry.Foundry(authorizers=auths)
     q = "Elwood"
     ds = f.search(q)
 
-    assert isinstance(ds, pd.DataFrame)
+    assert isinstance(ds, list)
     assert len(ds) > 0
-    assert ds.iloc[0]['name'] is not None
-    assert ds.iloc[0]['source_id'] is not None
-    assert ds.iloc[0]['year'] is not None
+    
+    # assert ds.iloc[0]['name'] is not None
+    assert ds[0].dc["titles"][0]["title"] is not None
+    
+    # assert ds.iloc[0]['source_id'] is not None
+    assert ds[0].name is not None
+
+    # assert ds.iloc[0]['year'] is not None
+    assert ds[0].dc.get("publicationYear", None) is not None
+
+
+def test_search_limit():
+    f = foundry.Foundry(authorizers=auths)
+    q = "atom"
+    ds = f.search(q, limit=100)
+
+    assert isinstance(ds, list)
+    assert len(ds) == 25
+    
+    # assert ds.iloc[0]['name'] is not None
+    assert ds[0].dc["titles"][0]["title"] is not None
+    
+    # assert ds.iloc[0]['source_id'] is not None
+    assert ds[0].name is not None
+
+    # assert ds.iloc[0]['year'] is not None
+    assert ds[0].dc.get("publicationYear", None) is not None
 
 
 def test_metadata_pull():
-    f = Foundry(test_dataset, download=False, authorizers=auths)
-    assert f.dc["titles"][0]["title"] == expected_title
+    f = foundry.Foundry(download=False, authorizers=auths)
+    dataset = f.search(test_dataset)
+    assert dataset[0].dc["titles"][0]["title"] == expected_title
 
 
[…]

[…]
