+++ foundry/foundry_dataset.py
@@ -3,7 +3,6 @@ import logging
 import os
 
 from json2table import convert
-import pandas as pd
 import numpy as np
 
 from foundry.models import FoundrySchema
