+++ docs/foundry.md
@@ -1,0 +1,24 @@
+<!-- markdownlint-disable -->
+
+<a href="https://github.com/MLMI2-CSSI/foundry/tree/main/foundry/__init__.py#L0"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+
+# <kbd>module</kbd> `foundry`
+
+
+
+
+**Global Variables**
+---------------
+- **auth**
+- **utils**
+- **models**
+- **https_download**
+- **https_upload**
+- **foundry**
+
+
+
+
+---
+
+_This file was automatically generated via [lazydocs](https://github.com/ml-tooling/lazydocs)._
