### 2023-11-14 269c9fc
Implementing new search() function (and refactor) (#408)
* update to 0.6.0 for HTTPS pub

* Upload Foundry class load() function to default download using https (#340)

* Update setup.py

Fix version number for pyPI deploy

* Update setup.py version for pyPI

* Update requirements.txt to latest DLHub SDK

This is needed to require upgrade of DLHub SDK for Foundry users when they upgrade Foundry.

* Update version to 0.6.3

* incorporate load() to foundry.__init__()

* automating api documentation using github action (#342)

* CI: Automated documentation build

* Removing remnants of XTract

* 
[…]
