+++ foundry/foundry_dataset.py
@@ -1,0 +1,221 @@
+import json
+import logging
+import os
+
+from json2table import convert
+import numpy as np
+
+from foundry.models import FoundrySchema
+from pydantic import ValidationError
+
+
+logger = logging.getLogger(__name__)
+
+
+class FoundryDataset(FoundrySchema):
+    """Representation of an individual dataset.
+        Provides access to metadata as well as functions to
+        instantiate data into memory in different formats.
+
+        Desired functions:
+            - Get as pandas
+            - Get as tensorflow dataset
+            - Get as pytorch dataset
+            - Get file list
+            - Set metadata
+            - Attach datafiles
+            - Validate against schema
+            - Get citation
+        """
+
+    def __init__(self, source_id: str, datacite_entry: dict, foundry_schema: FoundrySchema):
+        self.source_id = source_id
+        self.dc = datacite_entry
+        self.foundry_schema = foundry_schema
+
+    def to_torch(self, split: str = None):
+        """Convert Foundry Dataset to a PyTorch Dataset
+
+        Arguments:
+            split (string): Split to create PyTorch Dataset on.
+                    **Default:** ``None``
+
+        Returns: (TorchDataset) PyTorch Dataset of all the data from the specified split
+
+        """
+
+        self.download_if_not_downloaded()
+
+        from foundry.loaders.torch_wrapper import TorchDataset
+
+        inputs, targets = self._get_inputs_targets(split)
+        return TorchDataset(inputs, targets)
+
+    def to_tensorflow(self, split: str = None):
+        """Convert Foundry Dataset to a Tensorflow Sequence
+
+        Arguments:
+            split (string): Split to create Tensorflow Sequence on.
+                    **Default:** ``None``
+
+        Returns: (TensorflowSequence) Tensorflow Sequence of all the data from the specified split
+
+        """
+        from foundry.loaders.tf_wrapper import TensorflowSequence
+
+        inputs, targets = self._get_inputs_targets(split)
+        return TensorflowSequence(inputs, targets)
+
+    def download_if_not_downloaded(self):
+        ...
+
+    def load_data(self, source_id=None, globus=True, as_hdf5=False, splits=[]):
+        """Load in the data associated with the prescribed dataset
+
+        Tabular Data Type: Data are arranged in a standard data frame
+        stored in self.dataframe_file. The contents are read, and
+
+        File Data Type: <<Add desc>>
+
+        For more complicated data structures, users should
+        subclass FoundryDataset and override the load_data function
+
+        Args:
+           inputs (list): List of strings for input columns
+           targets (list): List of strings for output columns
+           source_id (string): Relative path to the source file
+           as_hdf5 (bool): If True and dataset is in hdf5 format, keep data in hdf5 format
+           splits (list): Labels of splits to be loaded
+
+        Returns:
+             (dict): a labeled dictionary of tuples
+        """
+        data = {}
+
+        # Handle splits if they exist. Return as a labeled dictionary of tuples
+        try:
+            if self.dataset.splits:
+                if not splits:
+                    for split in self.dataset.splits:
+                        data[split.label] = self._load_data(file=split.path, source_id=source_id, globus=globus,
+                                                            as_hdf5=as_hdf5)
+                else:
+                    for split in self.dataset.splits:
+                        if split.label in splits:
+                            splits.remove(split.label)
+                            data[split.label] = self._load_data(file=split.path, source_id=source_id, globus=globus,
+                                                                as_hdf5=as_hdf5)
+                    if len(splits) > 0:
+                        raise ValueError(f'The split(s) {splits} were not found in the dataset!')
+                return data
+            else:
[…]
