+++ tests/test_foundry.py
@@ -6,11 +6,12 @@ from filecmp import cmp
 from datetime import datetime
 from math import floor
 import numpy as np
+from pydantic import ValidationError
 import requests
 import mdf_toolbox
 import pandas as pd
 from mdf_forge import Forge
-from foundry import Foundry
+from foundry import foundry
 from foundry.auth import PubAuths
 from foundry.https_upload import upload_to_endpoint
 from dlhub_sdk import DLHubClient
@@ -134,6 +135,66 @@ pub_test_metadata = {
     'n_items': 1000
 }
 
+
+pub_test_invalid_metadata = {
+    "keys": [
+        {
+            "key": ["sepal length (cm)"],
+            "type": "input",
+            "units": "cm",
+            "description": 10
+        },
+        {
+            "key": ["sepal width (cm)"],
+            "type": "input",
+            "units": "cm",
+            "description": "sepal width in unit(cm)"
+        },
+        {
+            "key": ["petal length (cm)"],
+            "type": "input",
+            "units": "cm",
+            "description": "petal length in unit(cm)"
+        },
+        {
+            "key": ["petal width (cm)"],
+            "type": "input",
+            "units": "cm",
+            "description": "petal width in unit(cm)"
+        },
+        {
+            "key": ["y"],
+            "type": "output",
+            "units": "",
+            "description": "flower type",
+            "classes": [
+                {
+                    "label": "0",
+                    "name": "setosa"
+                },
+                {
+                    "label": "1",
+                    "name": "versicolor"
+                },
+                {
+                    "label": "2",
+                    "name": "virginica"
+                }
+            ]
+        }
+    ],
+    'splits': [
+        {'label': 'train', 'path': 'train.json', 'type': 'train'},
+        {'label': 'test', 'path': 'test.json', 'type': 'test'}
+    ],
+    "short_name": "example_AS_iris_test_{:.0f}".format(datetime.now().timestamp()),
+    "data_type": "tabular",
+    'task_type': ['unsupervised', 'generative'],
+    'domain': ['materials science', 'chemistry'],
+    'n_items': 1000
+}
+
+
 # Globus endpoint for '_iris_dev' for test publication
 pub_test_data_source = "https://app.globus.org/file-manager?origin_id=e38ee745-6d04-11e5-ba46-22000b92c6ec&origin_path=%2Ffoundry-test%2Firis-dev%2F"
 
@@ -146,64 +207,86 @@ def _delete_test_data(foundry_obj):
 
 
 def test_foundry_init():
-    f = Foundry(authorizers=auths)
+    f = foundry.Foundry(authorizers=auths)
     assert isinstance(f.forge_client, Forge)
     assert isinstance(f.connect_client, MDFConnectClient)
 
     if not is_gha:
         assert isinstance(f.dlhub_client, DLHubClient)
 
-        f2 = Foundry(authorizers=auths, no_browser=False, no_local_server=True)
+        f2 = foundry.Foundry(download=False, authorizers=auths, no_browser=False, no_local_server=True)
         assert isinstance(f2.dlhub_client, DLHubClient)
         assert isinstance(f2.forge_client, Forge)
         assert isinstance(f2.connect_client, MDFConnectClient)
 
-        f3 = Foundry(authorizers=auths, no_browser=True, no_local_server=False)
+        f3 = foundry.Foundry(download=False, authorizers=auths, no_browser=True, no_local_server=False)
         assert isinstance(f3.dlhub_client, DLHubClient)
[…]
