### 2023-11-14 8065b6f
CI: Automated documentation build
