+++ foundry/__init__.py
@@ -1,4 +1,1 @@
-from .foundry import Foundry  # noqa F401 (import unused)
-from . import models  # noqa F401 (import unused)
-from . import https_download  # noqa F401 (import unused)
-from . import https_upload  # noqa F401 (import unused)
+from .foundry import Foundry
