+++ foundry/foundry.py
@@ -16,11 +16,13 @@ from .utils import _read_csv, _read_json, _read_excel
 
 from foundry.models import (
     FoundrySchema,
-    FoundryDataset,
     FoundryBase
 )
 
-from foundry.foundry_cache import FoundryCache
+from foundry.foundry_dataset import FoundryDataset
+
+# from foundry import FOUNDRY_CACHE
+# from foundry.foundry_cache import FoundryCache
 
 from foundry.https_upload import upload_to_endpoint
 
@@ -44,11 +46,15 @@ class Foundry(FoundryBase):
     index = ""
     auths: Any
 
-    def __init__(
-            self, no_browser=False, no_local_server=False, index="mdf", authorizers=None,
-            globus=True, verbose=False, interval=10,
-            **data
-    ):
+    def __init__(self, 
+                 no_browser=False, 
+                 no_local_server=False, 
+                 index="mdf", 
+                 authorizers=None,
+                 globus=True, 
+                 verbose=False, 
+                 interval=10,
+                 **data):
         """Initialize a Foundry client
         Args:
             no_browser (bool):  Whether to open the browser for the Globus Auth URL.
@@ -68,7 +74,9 @@ class Foundry(FoundryBase):
         super().__init__(**data)
         self.index = index
         self.auths = None
-        self.cache = FoundryCache()
+        self.globus = globus
+        self.verbose = verbose
+        self.interval = interval
 
         if authorizers:
             self.auths = authorizers
@@ -189,11 +197,18 @@ class Foundry(FoundryBase):
             FoundryDataset: a FoundryDataset object created from the metadata
         """
         try:
-            schema = FoundrySchema(**metadata['projects']['foundry'])
+            foundry_schema = FoundrySchema(**metadata['projects']['foundry'])
             dc = metadata['dc']
             name = metadata['mdf']['source_id']
 
-            ds = FoundryDataset(**{'name': name, 'schema': schema, 'dc': dc})
+            ds = FoundryDataset(**{'dataset_name': name, 
+                                   'foundry_schema': foundry_schema,
+                                   'transfer_client': self.auths["transfer"],
+                                   'datacite_entry': dc, 
+                                   'globus': self.globus, 
+                                   'interval': self.interval, 
+                                   'verbose': self.verbose,
+                                   'forge_client': self.forge_client})
 
             return ds
 
@@ -257,10 +272,15 @@ class Foundry(FoundryBase):
         metadata = forge.search(q, advanced=True, limit=limit)
         return metadata
 
-    def publish_dataset(
-            self, foundry_metadata: Dict[str, Any], title: str, authors: List[str], https_data_path: str = None,
-            globus_data_source: str = None, update: bool = False, publication_year: int = None, test: bool = False,
-            **kwargs: Dict[str, Any],) -> Dict[str, Any]:
+    def publish_dataset(self, 
+                        foundry_metadata: Dict[str, Any], 
+                        title: str, authors: List[str], 
+                        https_data_path: str = None,
+                        globus_data_source: str = None, 
+                        update: bool = False, 
+                        publication_year: int = None, 
+                        test: bool = False,
+                        **kwargs: Dict[str, Any],) -> Dict[str, Any]:
         """Submit a dataset for publication; can choose to submit via HTTPS using `https_data_path` or via Globus
             Transfer using the `globus_data_source` argument. Only one upload method may be specified.
         Args:
@@ -404,97 +424,3 @@ class Foundry(FoundryBase):
     #     # return self.dlhub_client.get_task_status(res)
     #     pass
 
-    def get_keys(self, type=None, as_object=False):
-        """Get keys for a Foundry dataset
-
-        Arguments:
-            type (str): The type of key to be returned e.g., "input", "target"
-            as_object (bool): When ``False``, will return a list of keys in as strings
[…]
