+++ tests/test_foundry.py
@@ -1,10 +1,13 @@
+import builtins
 import json
+import mock
 import os
 import shutil
 import pytest
 from filecmp import cmp
 from datetime import datetime
 from math import floor
+
 import numpy as np
 from pydantic import ValidationError
 import requests
@@ -200,8 +203,8 @@ pub_test_data_source = "https://app.globus.org/file-manager?origin_id=e38ee745-6
 
 
 # Quick function to delete any downloaded test data
-def _delete_test_data(foundry_obj):
-    path = os.path.join(foundry_obj.config.local_cache_dir, test_dataset)
+def _delete_test_data(dataset):
+    path = os.path.join(dataset._foundry_cache.local_cache_dir, test_dataset)
     if os.path.isdir(path):
         shutil.rmtree(path)
 
@@ -244,7 +247,7 @@ def test_search():
     assert ds[0].dc["titles"][0]["title"] is not None
     
     # assert ds.iloc[0]['source_id'] is not None
-    assert ds[0].name is not None
+    assert ds[0].dataset_name is not None
 
     # assert ds.iloc[0]['year'] is not None
     assert ds[0].dc.get("publicationYear", None) is not None
@@ -252,17 +255,16 @@ def test_search():
 
 def test_search_limit():
     f = foundry.Foundry(authorizers=auths)
-    q = "atom"
-    ds = f.search(q, limit=100)
+    ds = f.search(limit=10)
 
     assert isinstance(ds, list)
-    assert len(ds) < 101
+    assert len(ds) == 10
     
     # assert ds.iloc[0]['name'] is not None
     assert ds[0].dc["titles"][0]["title"] is not None
     
     # assert ds.iloc[0]['source_id'] is not None
-    assert ds[0].name is not None
+    assert ds[0].dataset_name is not None
 
     # assert ds.iloc[0]['year'] is not None
     assert ds[0].dc.get("publicationYear", None) is not None
@@ -274,27 +276,42 @@ def test_metadata_pull():
     assert dataset[0].dc["titles"][0]["title"] == expected_title
 
 
-@pytest.mark.skip(reason='Omitting testing beyond search functionality until next story')
-def test_download_https():
-    f = Foundry(test_dataset, download=True, globus=False, authorizers=auths)
-    _delete_test_data(f)
-
-    assert f.dc["titles"][0]["title"] == expected_title
-    _delete_test_data(f)
-
-
-@pytest.mark.skip(reason='Omitting testing beyond search functionality until next story')
-def test_dataframe_load():
-    f = Foundry(test_dataset, download=True, globus=False, authorizers=auths)
-
-    res = f.load_data()
+def test_download_globus():
+    f = foundry.Foundry(globus=True, authorizers=auths)
+    dataset = f.search(test_dataset)[0]
+    res = dataset.get_as_dict()
     X, y = res['train']
 
     assert len(X) > 1
     assert isinstance(X, pd.DataFrame)
     assert len(y) > 1
     assert isinstance(y, pd.DataFrame)
-    _delete_test_data(f)
+    _delete_test_data(dataset)
+
+
+@pytest.mark.skip(reason='Not sure why, but http download is bringing up a globus auth page instead of a datafile')
+def test_download_https():
+    f = foundry.Foundry(globus=False, authorizers=auths)
+    dataset = f.search(test_dataset)[0]
+    res = dataset.get_as_dict()
+    X, y = res['train']
+
+    assert len(X) > 1
+    assert isinstance(X, pd.DataFrame)
+    assert len(y) > 1
+    assert isinstance(y, pd.DataFrame)
+    _delete_test_data(dataset)
+
+
[…]
