### 2023-12-06 914de73
updated docstrings and examples, added extended pandas dataframe class for display and access of datasets
