+++ tests/test_foundry.py
@@ -240,45 +240,79 @@ def test_search():
     q = "Elwood"
     ds = f.search(q)
 
+    assert isinstance(ds, pd.DataFrame)
+    assert len(ds) > 0
+    
+    dataset = ds.iloc[0].FoundryDataset
+
+    # assert ds.iloc[0]['name'] is not None
+    assert dataset.dc["titles"][0]["title"] is not None
+    
+    # assert ds.iloc[0]['source_id'] is not None
+    assert dataset.dataset_name is not None
+
+    # assert ds.iloc[0]['year'] is not None
+    assert dataset.dc.get("publicationYear", None) is not None
+
+
+def test_search_as_list():
+    f = foundry.Foundry(authorizers=auths)
+    q = "Elwood"
+    ds = f.search(q, as_list=True)
+
     assert isinstance(ds, list)
     assert len(ds) > 0
     
-    # assert ds.iloc[0]['name'] is not None
-    assert ds[0].dc["titles"][0]["title"] is not None
-    
-    # assert ds.iloc[0]['source_id'] is not None
-    assert ds[0].dataset_name is not None
+    dataset = ds[0]
 
-    # assert ds.iloc[0]['year'] is not None
-    assert ds[0].dc.get("publicationYear", None) is not None
+    # assert ds[0]['name'] is not None
+    assert dataset.dc["titles"][0]["title"] is not None
+    
+    # assert ds[0]['source_id'] is not None
+    assert dataset.dataset_name is not None
+
+    # assert ds[0]['year'] is not None
+    assert dataset.dc.get("publicationYear", None) is not None
 
 
 def test_search_limit():
     f = foundry.Foundry(authorizers=auths)
     ds = f.search(limit=10)
 
-    assert isinstance(ds, list)
+    assert isinstance(ds, pd.DataFrame)
     assert len(ds) == 10
     
+    dataset = ds.iloc[0].FoundryDataset
+
     # assert ds.iloc[0]['name'] is not None
-    assert ds[0].dc["titles"][0]["title"] is not None
+    assert dataset.dc["titles"][0]["title"] is not None
     
     # assert ds.iloc[0]['source_id'] is not None
-    assert ds[0].dataset_name is not None
+    assert dataset.dataset_name is not None
 
     # assert ds.iloc[0]['year'] is not None
-    assert ds[0].dc.get("publicationYear", None) is not None
+    assert dataset.dc.get("publicationYear", None) is not None
+
+
+def test_search_no_results():
+    f = foundry.Foundry()
+
+    with pytest.raises(Exception) as exc_info:
+        f.search('chewbacca')
+
+    err = exc_info.value
+    assert hasattr(err, '__cause__')
 
 
 def test_metadata_pull():
     f = foundry.Foundry(download=False, authorizers=auths)
-    dataset = f.search(test_dataset)
-    assert dataset[0].dc["titles"][0]["title"] == expected_title
+    dataset = f.search(test_dataset).iloc[0].FoundryDataset
+    assert dataset.dc["titles"][0]["title"] == expected_title
 
 
 def test_download_globus():
     f = foundry.Foundry(globus=True, authorizers=auths)
-    dataset = f.search(test_dataset)[0]
+    dataset = f.search(test_dataset).iloc[0].FoundryDataset
     res = dataset.get_as_dict()
     X, y = res['train']
 
@@ -292,21 +326,14 @@ def test_download_globus():
 @pytest.mark.skip(reason='Not sure why, but http download is bringing up a globus auth page instead of a datafile')
 def test_download_https():
     f = foundry.Foundry(globus=False, authorizers=auths)
-    dataset = f.search(test_dataset)[0]
-    res = dataset.get_as_dict()
-    X, y = res['train']
-
-    assert len(X) > 1
[…]
