+++ foundry/foundry.py
@@ -1,9 +1,7 @@
-import h5py
 import mdf_toolbox
 import pandas as pd
 from typing import Any, Dict, List
 import logging
-import os
 
 from mdf_connect_client import MDFConnectClient
 from mdf_forge import Forge
@@ -11,8 +9,7 @@ from dlhub_sdk import DLHubClient
 from globus_sdk import AuthClient
 
 from .auth import PubAuths
-from .utils import is_pandas_pytable, is_doi
-from .utils import _read_csv, _read_json, _read_excel
+from .utils import is_doi
 
 from foundry.models import (
     FoundrySchema,
@@ -30,11 +27,64 @@ logging.basicConfig(format='%(levelname)s: %(message)s', level=logging.WARNING)
 logger = logging.getLogger(__name__)
 
 
+class HiddenColumnDataFrame(pd.DataFrame):
+    """
+    A subclass of pd.DataFrame that supports hiding a specific column.
+
+    Parameters:
+    *args: positional arguments
+        Positional arguments passed to the parent class constructor.
+    hidden_column: str, optional
+        The name of the column to be hidden.
+    **kwargs: keyword arguments
+        Keyword arguments passed to the parent class constructor.
+
+    Attributes:
+    hidden_column: str or None
+        The name of the hidden column.
+
+    Methods:
+    _repr_html_():
+        Overrides the _repr_html_ method of the parent class to hide the specified column in the HTML representation.
+    get_dataset_by_name(dataset_name):
+        Returns the FoundryDataset associated with the given dataset name. Can also handle a DOI.
+    get_dataset_by_doi(doi):
+        Returns the FoundryDataset associated with the given DOI.
+
+    """
+
+    def __init__(self, *args, hidden_column=None, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.hidden_column = hidden_column
+
+    def _repr_html_(self):
+        if self.hidden_column is not None and self.hidden_column in self.columns:
+            return self.drop(self.hidden_column, axis=1)._repr_html_()
+
+    def get_dataset_by_name(self, dataset_name):
+        if is_doi(dataset_name):
+            return self.get_dataset_by_doi(dataset_name)
+        else:
+            return self[self.dataset_name == dataset_name]['FoundryDataset'].iloc[0]
+
+    def get_dataset_by_doi(self, doi):
+        return self[self.DOI == doi]['FoundryDataset'].iloc[0]
+
+
 class Foundry(FoundryBase):
     """Foundry Client Base Class
-    TODO:
-    -------
-    Add Docstring
+
+    This class represents a client for interacting with the Foundry service. It provides methods for searching and
+    accessing datasets, as well as publishing new datasets.
+
+    Attributes:
+        dlhub_client (Any): The DLHub client.
+        forge_client (Any): The Forge client.
+        connect_client (Any): The MDF Connect client.
+        transfer_client (Any): The Globus transfer client.
+        auth_client (Any): The authentication client.
+        index (str): The index to use for search and data publication.
+        auths (Any): The authorizers used for authentication.
 
     """
 
@@ -46,30 +96,29 @@ class Foundry(FoundryBase):
     index = ""
     auths: Any
 
-    def __init__(self, 
-                 no_browser=False, 
-                 no_local_server=False, 
-                 index="mdf", 
+    def __init__(self,
+                 no_browser=False,
+                 no_local_server=False,
+                 index="mdf",
                  authorizers=None,
-                 globus=True, 
-                 verbose=False, 
+                 globus=True,
+                 verbose=False,
                  interval=10,
                  **data):
         """Initialize a Foundry client
+
         Args:
-            no_browser (bool):  Whether to open the browser for the Globus Auth URL.
[…]
