### 2023-12-06 493453f
updated docstrings and examples, added extended pandas dataframe class for display and access of datasets
