+++ foundry/foundry_cache.py
@@ -2,14 +2,16 @@ import logging
 import os
 
 from concurrent.futures import ThreadPoolExecutor, as_completed
+import h5py
 from mdf_forge import Forge
+import pandas as pd
 import shutil
 from tqdm.auto import tqdm
 from typing import List, Any
 
 from foundry.https_download import recursive_ls, download_file
 from foundry.models import FoundrySplit, FoundrySchema
-from .utils import _read_csv, _read_json, _read_excel
+from .utils import _read_csv, _read_json, _read_excel, is_pandas_pytable
 
 logger = logging.getLogger(__name__)
 
@@ -17,32 +19,43 @@ logger = logging.getLogger(__name__)
 class FoundryCache():
     """The FoundryCache manages the local storage of FoundryDataset objects"""
 
-    def __init__(self, 
+    def __init__(self,
                  forge_client: Forge,
                  transfer_client: Any):
+        """
+        Initializes a FoundryCache object.
 
+        Args:
+            forge_client (Forge): The Forge client object.
+            transfer_client (Any): The transfer client object.
+        """
         self.local_cache_dir = os.environ.get("FOUNDRY_LOCAL_CACHE_DIR", "./data")
         self.forge_client = forge_client
         self.transfer_client = transfer_client
 
-    def download_to_cache(self, 
+    def download_to_cache(self,
                           dataset_name: str,
-                          splits: List[FoundrySplit] = None, 
-                          globus: bool = False, 
-                          interval: int = 10, 
-                          parallel_https: int = 4, 
+                          splits: List[FoundrySplit] = None,
+                          globus: bool = False,
+                          interval: int = 10,
+                          parallel_https: int = 4,
                           verbose: bool = False,
-                          transfer_client = None):
-        """Downloads the data from source to local storage
-        
-        Args:
-            dataset_name (str): name of dataset (equivalrnt to source_id in MDF)
-            splits List[FoundrySplit]: list of splits in the dataset
-            globus (bool): if True, use Globus to download the data else try HTTPS
-            interval (int): How often to wait before checking Globus transfer status
-            parallel_https (int): Number of files to download in parallel if using HTTPS
-            verbose (bool): Produce more debug messages to screen
+                          transfer_client=None):
 
+        """
+        Downloads the data from source to local storage.
+
+        Args:
+            dataset_name (str): Name of the dataset (equivalent to source_id in MDF).
+            splits (List[FoundrySplit], optional): List of splits in the dataset. Defaults to None.
+            globus (bool, optional): If True, use Globus to download the data; otherwise, try HTTPS. Defaults to False.
+            interval (int, optional): How often to wait before checking Globus transfer status. Defaults to 10.
+            parallel_https (int, optional): Number of files to download in parallel if using HTTPS. Defaults to 4.
+            verbose (bool, optional): Produce more debug messages to screen. Defaults to False.
+            transfer_client (Any, optional): The transfer client object. Defaults to None.
+
+        Returns:
+            FoundryCache: The FoundryCache object.
         """
         if not self.validate_local_dataset_storage(dataset_name, splits):
             if globus:
@@ -54,15 +67,15 @@ class FoundryCache():
 
         return self
 
-    def download_via_globus(self, 
-                            dataset_name: str, 
+    def download_via_globus(self,
+                            dataset_name: str,
                             interval: int):
-        """ Downloads selected dataset over globus
+        """
+        Downloads selected dataset over Globus.
 
         Args:
-            dataset_name (str): name of dataset (equivalent to source_id in MDF)
-            interval (int): How often to wait before checking Globus transfer status
-
+            dataset_name (str): Name of the dataset (equivalent to source_id in MDF).
+            interval (int): How often to wait before checking Globus transfer status.
         """
         # query for mdf data representation
[…]
