### 2023-12-07 fb387b4
cleaning up a few loose ends
