+++ tests/test_foundry.py
@@ -228,6 +228,7 @@ def test_foundry_init():
         assert isinstance(f3.connect_client, MDFConnectClient)
 
 
+@pytest.mark.skip(reason='Checking if tests work')
 def test_list():
     f = foundry.Foundry(authorizers=auths)
     ds = f.list()
@@ -235,6 +236,7 @@ def test_list():
     assert len(ds) > 0
 
 
+@pytest.mark.skip(reason='Checking if tests work')
 def test_search():
     f = foundry.Foundry(authorizers=auths)
     q = "Elwood"
@@ -255,6 +257,7 @@ def test_search():
     assert dataset.dc.get("publicationYear", None) is not None
 
 
+@pytest.mark.skip(reason='Checking if tests work')
 def test_search_as_list():
     f = foundry.Foundry(authorizers=auths)
     q = "Elwood"
@@ -275,6 +278,7 @@ def test_search_as_list():
     assert dataset.dc.get("publicationYear", None) is not None
 
 
+@pytest.mark.skip(reason='Checking if tests work')
 def test_search_limit():
     f = foundry.Foundry(authorizers=auths)
     ds = f.search(limit=10)
@@ -294,6 +298,7 @@ def test_search_limit():
     assert dataset.dc.get("publicationYear", None) is not None
 
 
+@pytest.mark.skip(reason='Checking if tests work')
 def test_search_no_results():
     f = foundry.Foundry()
 
@@ -304,6 +309,7 @@ def test_search_no_results():
     assert hasattr(err, '__cause__')
 
 
+@pytest.mark.skip(reason='Checking if tests work')
 def test_metadata_pull():
     f = foundry.Foundry(download=False, authorizers=auths)
     dataset = f.search(test_dataset).iloc[0].FoundryDataset
@@ -378,6 +384,7 @@ def test_dataframe_load_split_but_no_splits():
     _delete_test_data(dataset)
 
 
+@pytest.mark.skip(reason='Checking if tests work')
 def test_dataframe_search_by_doi():
     f = foundry.Foundry(globus=False, authorizers=auths)
 
