### 2023-12-20 85e74c7
first pass at docstring updates
