+++ foundry/foundry.py
@@ -35,10 +35,9 @@ logger = logging.getLogger(__name__)
 
 class Foundry(FoundryBase):
     """Foundry Client Base Class
-    TODO:
-    -------
-    Add Docstring
 
+    Foundry object used for all interactions with Foundry datasets and models. Interfaces with MDF Connect Client,
+        Globus Compute, Globus Auth, Globus Transfer, Globus Search, DLHub, and relevant Globus Endpoints
     """
 
     dlhub_client: Any
@@ -55,6 +54,7 @@ class Foundry(FoundryBase):
             **data
     ):
         """Initialize a Foundry client
+
         Args:
             name (str): Name of the foundry dataset. If not supplied, metadata will not be loaded into
                     the Foundry object
@@ -166,6 +166,7 @@ class Foundry(FoundryBase):
 
     def _load(self, name, download=True, globus=True, verbose=False, metadata=None, authorizers=None, interval=None):
         """Load the metadata for a Foundry dataset into the client
+
         Args:
             name (str): Name of the foundry dataset
             download (bool): If True, download the data associated with the package (default is True)
@@ -228,12 +229,14 @@ class Foundry(FoundryBase):
 
     def search(self, q=None, limit=None):
         """Search available Foundry datasets
-        q (str): query string to match
-        limit (int): maximum number of results to return
 
-        Returns
-        -------
-            (pandas.DataFrame): DataFrame with summary list of Foundry data packages including name, title, publication year, and DOI
+        Args:
+            q (str): query string to match
+            limit (int): maximum number of results to return
+
+        Returns:
+            (pandas.DataFrame): DataFrame with summary list of Foundry data packages including name, title, publication
+                year, and DOI
         """
         if not q:
             q = None
@@ -258,22 +261,24 @@ class Foundry(FoundryBase):
 
     def list(self):
         """List available Foundry datasets
-        Returns
-        -------
-            (pandas.DataFrame): DataFrame with summary list of Foundry datasets including name, title, publication year, and DOI
+
+        Returns:
+            (pandas.DataFrame): DataFrame with summary list of Foundry datasets including name, title, publication
+                year, and DOI
         """
         return self.search()
 
     def run(self, name, inputs, funcx_endpoint=None, **kwargs):
-        """Run a model on data
+        """Run a model on inputted data
 
         Args:
-           name (str): DLHub model name
-           inputs: Data to send to DLHub as inputs (should be JSON serializable)
-           funcx_endpoint (optional): UUID for the funcx endpoint to run the model on, if not the default (eg River)
+            name (str): DLHub model name
+            inputs: Data to send to DLHub as inputs (should be JSON serializable, example types include dict, list,
+                np.ndarray, etc)
+            funcx_endpoint (str) (optional): UUID for the funcx endpoint to run the model on, if not the default (eg River)
 
         Returns:
-             Returns results after invocation via the DLHub service
+             Results after invocation via the DLHub service
         """
         if funcx_endpoint is not None:
             self.dlhub_client.fx_endpoint = funcx_endpoint
@@ -291,14 +296,13 @@ class Foundry(FoundryBase):
         subclass Foundry and override the load_data function
 
         Args:
-           inputs (list): List of strings for input columns
-           targets (list): List of strings for output columns
-           source_id (string): Relative path to the source file
+           source_id (str): Name of the dataset in MDF/Foundry index (``source_name`` + version information)
+           globus (bool): If True, download using Globus, otherwise, HTTPS
            as_hdf5 (bool): If True and dataset is in hdf5 format, keep data in hdf5 format
            splits (list): Labels of splits to be loaded
 
         Returns:
-             (dict): a labeled dictionary of tuples
+            data (dict): a labeled dictionary of tuples
         """
         data = {}
 
[…]
