### 2024-01-02 0cddec9
GITBOOK-109: Adding documentation of publishing links with metadata
