### 2024-01-02 MLMI2-CSSI/foundry#375 @ascourtas
Based on clicking around, it appears that the links in the [actual GitBook documentation](https://ai-materials-and-chemistry.gitbook.io/foundry/) are not broken, but rather the `docs` hosted in the repo are significantly out of synch with the GitBook docs, thus making it confusing for the user. GitBook used to automatically synch, but in one of their updates that stopped working.

I'm going to reframe this issue to do the following:

- [ ] Re-synch GitBook with GitHub to update the `docs` md files
- [ ] Double check to make sure the GitBook links work
- [ ] Ensure GitBook docs reflect current functionality

In future work, we may want to consider moving away from GitBook
