+++ docs/README.md
@@ -1,43 +1,26 @@
-# Getting started with Foundry
+# Introduction and Installation
 
-![](.gitbook/assets/foundry-purple%20%283%29.png)
+<figure><img src=".gitbook/assets/foundry-logo-new.png" alt=""><figcaption></figcaption></figure>
 
 ## What is Foundry?
 
-Foundry is a Python package that simplifies the discovery and usage of machine-learning ready datasets in materials science and chemistry. Foundry provides software tools that make it easy to load these datasets and work with them in local or cloud environments. Further, Foundry provides a dataset specification, and defined curation flows, that allow users to create new datasets for the community to use through this same interface.
+Foundry is a Python package that simplifies the discovery and usage of machine-learning ready datasets and published models in materials science and chemistry. We provide software tools that make it easy to load datasets and work with them in local or cloud environments and to perform inference using published ML models.&#x20;
+
+**Capability Status:** Researchers can contribute by publishing new [datasets](publishing/publishing-datasets.md) and [models](under-development/publishing-models.md) through a self-service publication process. Our dataset capabilities are currently in **production** and our model capabilities are still **in development**.
 
 ## Installation
 
 Foundry can be installed on any operating system with Python with pip
 
-```text
+```
 pip install foundry-ml
 ```
 
+<table data-view="cards"><thead><tr><th></th><th></th><th></th><th data-hidden data-card-target data-type="content-ref"></th><th data-hidden data-card-cover data-type="files"></th></tr></thead><tbody><tr><td></td><td>Use Datasets</td><td></td><td><a href="loading-and-using/examples.md">examples.md</a></td><td><a href=".gitbook/assets/foundry-down.png">foundry-down.png</a></td></tr><tr><td></td><td>Publish Datasets</td><td></td><td><a href="publishing/publishing-datasets.md">publishing-datasets.md</a></td><td><a href=".gitbook/assets/foundry-up.png">foundry-up.png</a></td></tr></tbody></table>
+
 ### Globus
 
-Foundry uses the Globus platform for authentication, search, and to optimize some data transfer operations. Follow the steps below to get set up.
+We use the Globus platform for authentication and optimization of some data transfer operations. Follow the steps below to get set up.
 
-* [Create a free account.](https://app.globus.org) You can create a free account here with your institutional credentials or with free IDs \(GlobusID, Google, ORCID, etc\).
-* [Set up a Globus Connect Personal endpoint ](https://www.globus.org/globus-connect-personal)_**\(optional\)**_. While this step is optional, some Foundry capabilities will work more efficiently when using GCP.
-
-## Project Support
-
-This work was supported by the National Science Foundation under NSF Award Number: 1931306 "Collaborative Research: Framework: Machine Learning Materials Innovation Infrastructure".
-
-### Other Support
-
-Foundry brings together many components in the materials data ecosystem. Including MAST-ML, the Data and Learning Hub for Science \(DLHub\), and The Materials Data Facility \(MDF\).
-
-#### MAST-ML
-
-This work was supported by the National Science Foundation \(NSF\) SI2 award No. 1148011 and DMREF award number DMR-1332851
-
-#### The Data and Learning Hub for Science \(DLHub\)
-
-This material is based upon work supported by Laboratory Directed Research and Development \(LDRD\) funding from Argonne National Laboratory, provided by the Director, Office of Science, of the U.S. Department of Energy under Contract No. DE-AC02-06CH11357. [https://www.dlhub.org](https://www.dlhub.org)
-
-#### The Materials Data Facility
-
[…]
