+++ docs/publishing/describing-datasets.md
@@ -1,0 +1,181 @@
+---
+description: An introduction to how Foundry datasets and their metadata are defined
+---
+
+# Describing Datasets
+
+Foundry Datasets are comprised of two key components, [_**data**_](describing-datasets.md#data) and descriptive [_**metadata**_](describing-datasets.md#descriptive-metadata). In order to make the data easily consumable, _**data**_ (consisting of files) should be assembled following the supported structures. The _**metadata**_ description allows tracking of high level information (e.g.,  authors, associated institutions, licenses, data location), and also information on how to operate on the datasets (e.g., how to load the data, training/test splits).
+
+## **Data**
+
+### Example - Record-Based Data
+
+#### **Tabular Data**
+
+For tabular data, columns should represent the different keys of the data, and rows the individual records.
+
+{% hint style="info" %}
+Supported tabular data types currently include .json, .jsonl, and .csv
+{% endhint %}
+
+In this example, we showcase how to describe a JSON record-based dataset where each record is a valid JSON object in a JSON list or a line in a JSON line delimited file.
+
+| **feature\_1** | **feature\_2** | **material\_type** | band\_gap |
+| -------------- | -------------- | ------------------ | --------- |
+| 0.10           | 0.52           | 1                  | 1.40      |
+| 0.34           | 0.910          | 0                  | 0.73      |
+| ...            | ...            | ...                |           |
+
+For this example dataset, the `Key` Python object could be: &#x20;
+
+```python
+"keys": [{
+	"key": "feature_1",
+	"type": "input",
+	"units": None,
+	"description": "This is feature 1"
+}, {
+	"key": "feature_2",
+	"type": "input",
+	"units": None,
+	"description": "This is feature 2"
+}, {
+	"key": "material_type",
+	"type": "input",
+	"units": None,
+	"description": "This is the material type",
+	"labels": ["perovskite", "not perovskite"]
+} {
+	"key": "band_gap",
+	"type": "target",
+	"units": "eV",
+	"description": "This is the simulated band gap in eV"
+}]
+```
+
+#### Hierarchical Data
+
+Foundry also supports data from hierarchical data formats (e.g., [HDF5](https://www.h5py.org)). In this case, features and outputs can be represented with `/` notation. For example, if the features of a dataset are located in an array stored in `/data/arr1` and `/other_data/arr2` while the outputs are in `/data/labeled`, the `keys` list may contain the following `Key` objects:
+
+```javascript
+"keys": [{
+		"key": ["/data/arr1"],
+		"type": "input",
+		"description": "input, unlabeled images"
+	}, {
+		"key": ["/data/arr2"],
+		"type": "input",
+		"description": "This is an another array containing input data"
+	}, {
+		"key": ["/data/labeled"],
+		"type": "target",
+		"description": "target, labeled images"
+	}]
+}
+```
+
+and the dataset may be described in whole as:
+
+```python
+{
+	"short_name": "segmentation-dev",
+	"data_type": "hdf5",
+	"task_type": ["unsupervised", "segmentation"],
+	"domain": ["materials science", "chemistry"],
+	"n_items": 100,
+	"splits": [{
+		"type": "train",
+		"path": "foundry.hdf5",
+		"label": "train"
+	}],
+	"keys": [{
+		"key": ["/data/arr1"],
+		"type": "input",
+		"description": "input, unlabeled images"
+	}, {
+		"key": ["/data/arr2"],
+		"type": "input",
+		"description": "This is an another array containing input data"
+	}, {
+		"key": ["/data/labeled"],
+		"type": "target",
+		"description": "target, labeled images"
+	}]
+}
+```
+
+## Descriptive Metadata
+
+Metadata in Foundry comprehensively describe datasets using a combination of the DataCite metadata schema and our own Foundry schema.
+
[…]
