### 2024-01-03 MLMI2-CSSI/foundry#375 @ascourtas
The GitBook - GitHub Integration has been fixed (thank you @blaiszik) and the `/docs` folder now reflects that. From here on all changes made in our [documentation](https://ai-materials-and-chemistry.gitbook.io/foundry/) will be reflected in `/docs`. Changes can be viewed in [commit 0fa7f19](https://github.com/MLMI2-CSSI/foundry/commit/0fa7f1950543a14b756b5e3f6bc21c608acb91c9)

The links to testing best practices were indeed broken in the current documentation, and have now been fixed and deployed in [commit 68c8fbf](https://github.com/MLMI2-CSSI/foundry/commit/68c8fbfcda620a5b39b9f22b01b22d13dc2a6797)

The link to our documentation is currently available in the GitHub repo's README, but please let us know if you think it needs to be more clearly communicated.

I will close this issue now that's resolved -- please feel free to reopen it if this solution didn't address all parts of the issue
