### 2024-01-03 3eceb30
wrapping up testing of foundry_cache.py
