### 2024-01-04 MLMI2-CSSI/foundry#373 @ascourtas
Updated [Quickstart docs](https://github.com/MLMI2-CSSI/foundry#quick-start) to reflect the proper usage; documentation has been deployed. I'll close this issue until further notice.
