+++ tests/test_foundry_cache.py
@@ -1,0 +1,154 @@
+import os
+from pathlib import Path
+import pytest
+from unittest.mock import MagicMock
+
+from foundry.models import FoundrySplit, FoundrySchema, FoundryDatasetType, FoundryKey
+from foundry.foundry_cache import FoundryCache
+from foundry.models import FoundryKey, FoundrySchema, FoundryDatasetType
+import pandas as pd
+import pytest
+from unittest.mock import MagicMock
+from unittest.mock import patch
+import pandas as pd
+from unittest.mock import patch
+
+@pytest.fixture
+def mock_foundry_cache():
+    cache_dir = str(Path(__file__).parent) + '/test_data'
+    cache = FoundryCache(forge_client=MagicMock(), transfer_client=MagicMock(), local_cache_dir=cache_dir)
+    # cache.validate_local_dataset_storage = MagicMock(return_value=True)  # Mock the validation method
+    return cache
+
+
+@pytest.fixture
+def mock_nonexistent_foundry_cache():
+    cache_dir = str(Path(__file__).parent) + '/cheeseballs'
+    cache = FoundryCache(forge_client=MagicMock(), transfer_client=MagicMock(), local_cache_dir=cache_dir)
+    # cache.validate_local_dataset_storage = MagicMock(return_value=True)  # Mock the validation method
+    return cache
+
+
+def test_validate_local_dataset_storage_exists(mock_foundry_cache):
+    cache = mock_foundry_cache
+    dataset_name = "elwood_md_v1.2"
+    assert cache.validate_local_dataset_storage(dataset_name) == True
+
+
+def test_validate_local_dataset_storage_missing_files(mock_foundry_cache):
+    cache = mock_foundry_cache
+    dataset_name = "elwood_md_v1.2"
+    # Create a split with a missing file
+    splits = [
+        FoundrySplit(path="file1.csv", type="train"),
+        FoundrySplit(path="file2.csv", type="test"),
+        FoundrySplit(path="file3.csv", type="validation")
+    ]
+
+    assert cache.validate_local_dataset_storage(dataset_name, splits) is False
+
+
+def test_validate_local_dataset_storage_complete(mock_foundry_cache):
+    cache = mock_foundry_cache
+    dataset_name = "elwood_md_v1.2"
+
+    assert cache.validate_local_dataset_storage(dataset_name) is True
+
+
+def test_validate_local_dataset_storage_not_present(mock_nonexistent_foundry_cache):
+    cache = mock_nonexistent_foundry_cache
+    dataset_name = "test_dataset"
+
+    assert cache.validate_local_dataset_storage(dataset_name) is False
+
+
+@pytest.fixture
+def mock_tabular_foundry_schema():
+    # foundry_key_1 = FoundryKey(key=["column1", "column2"], type='input')
+    # foundry_key_2 = FoundryKey(key=["column2", "column4"], type='output')
+    foundry_key_1 = FoundryKey(key=['SMILES'], type='input', classes=None, description='Canonical SMILES string of molecule', filter=None, units='arb')
+    foundry_key_2 = FoundryKey(key=['E_coh (MPa)'], type='target', classes=None, description='Simulated cohesive energy (in MPa)', filter=None, units='MPa')
+    foundry_key_3 = FoundryKey(key=['T_g (K)'], type='target', classes=None, description='Simulated glass transition temperature (in Kelvin)', filter=None, units='Kelvin')
+    foundry_key_4 = FoundryKey(key=['R_gyr (A^2)'], type='target', classes=None, description='Simulated squared radius of gyration (in Angstroms^2)', filter=None, units='Angstrom^2')
+    foundry_key_5 = FoundryKey(key=['Densities (kg/m^3)'], type='target', classes=None, description='Simulated density (in kg/m^3)', filter=None, units='kg/m^3')
[…]
