### 2024-01-08 MLMI2-CSSI/foundry#419 @ascourtas
address remaining JOSS comments
In addition to the opened issues by marshallmcdonnell, review any remaining comments at https://github.com/openjournals/joss-reviews/issues/5467 that may have been missed.

# Tasks
- [x] Determine if comments remain
- [x] Address them accordingly
