### 2024-01-09 MLMI2-CSSI/foundry#419 @ascourtas
Reviewers confirmed there are no remaining changes to be made. Last steps:

- [ ] release changes
- [ ] Zip code
- [ ] Upload zip to Zenodo, with the following reqs:
  - [ ]     The title of the archive is the same as the JOSS paper title
  - [ ]     That the authors of the archive are the same as the JOSS paper authors
- [ ] Reply to review thread with Zenodo DOI
