### 2024-01-10 #5467 @ascourtas
Hello @Fei-Tao , the code changes have been released in [v0.7.2 ](https://github.com/MLMI2-CSSI/foundry/releases/tag/v0.7.2)(with an additional release of [v0.7.3](https://github.com/MLMI2-CSSI/foundry/releases/tag/v0.7.3) to complete the Zenodo integration).

The Zenodo archive has been created with the DOI [10.5281/zenodo.10480757](https://zenodo.org/doi/10.5281/zenodo.10480757) (associated with the latest release).

Please let me know if any updates or other content is required. Thanks!
