### 2024-01-12 74a304d
remove duplicate Globus paper reference
