+++ JOSS/paper.md
@@ -125,7 +125,7 @@ Foundry-ML focuses foremost on accessibility and reproducibility. \autoref{fig1}
 Foundry-ML is built upon a solid base.
 We have developed Foundry-ML using the Materials Data Facility (MDF) [@mdf-2019; @mdf-2016] and Globus services like Auth, Transfer, and Search. Foundry-ML users can upload large datasets (MDF supports multi-TB databases, with potentially millions of files), making them easy to share, use, and discover by the rest of the scientific community.  All datasets are made available through the Foundry-ML software, the Foundry-ML webapp and also via Globus endpoints that support both Globus and HTTPS access.
 
-Beyond just simplified data access, enhanced interpretability is a key feature of Foundry-ML. Foundry-ML datasets have required metadata (see \autoref{fig1}b) that are provided by the authors of each dataset. All metadata are stored in Globus Search [@globuspublish] to facilitate queries. To make these metadata easily usable by Foundry-ML users, query helpers are provided via the Foundry-ML Python client to perform common actions e.g., listing all datasets, selecting datasets by DOI, and more.
+Beyond just simplified data access, enhanced interpretability is a key feature of Foundry-ML. Foundry-ML datasets have required metadata (see \autoref{fig1}b) that are provided by the authors of each dataset. All metadata are stored in Globus Search [@chard2015globus] to facilitate queries. To make these metadata easily usable by Foundry-ML users, query helpers are provided via the Foundry-ML Python client to perform common actions e.g., listing all datasets, selecting datasets by DOI, and more.
 
 In addition to the Python software interface to each dataset, we have developed a prototype web interface (\autoref{fig2}) that lists all datasets with instructions on how to access them and key features of each dataset (e.g., number of entries, inputs, targets, type of data, tags, free text description). While the examples presented here come from the domains of materials science and chemistry, Foundry-ML is designed to be domain agnostic, and since similar problems exist in other domains, we expect these approaches to generalize. Generalizing to other domains will allow the same software and services to help solve similar problems across scientific domains.
 
