### 2024-01-12 cb1c5b5
Update paper.bib with DOIs and correct citations
