+++ JOSS/paper.bib
@@ -7,135 +7,139 @@
 }
 
 @article{mdf-2019,
-  title={A data ecosystem to support machine learning in materials science},
+  title={A Data Ecosystem to Support Machine Learning in Materials Science},
   author={Blaiszik, Ben and Ward, Logan and Schwarting, Marcus and Gaff, Jonathon and Chard, Ryan and Pike, Daniel and Chard, Kyle and Foster, Ian},
   journal={MRS Communications},
   volume={9},
   number={4},
   pages={1125--1133},
   year={2019},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={10.1557/mrc.2019.118}
 }
 
 @article{mdf-2016,
-  title={The materials data facility: data services to advance materials science research},
+  title={The Materials Data Facility: Data Services to Advance Materials Science Research},
   author={Blaiszik, Ben and Chard, Kyle and Pruyne, Jim and Ananthakrishnan, Rachana and Tuecke, Steven and Foster, Ian},
   journal={Jom},
   volume={68},
   number={8},
   pages={2045--2052},
   year={2016},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11837-016-2001-3}
 }
 
 @article{wei2022benchmark,
-  title={Benchmark tests of atom segmentation deep learning models with a consistent dataset},
+  title={Benchmark Tests of Atom Segmentation Deep Learning Models with a Consistent Dataset},
   author={Wei, Jingrui and Blaiszik, Ben and Scourtas, Aristana and Morgan, Dane and Voyles, Paul M},
   journal={arXiv preprint arXiv:2207.10173},
-  year={2022}
+  year={2022},
+  doi={10.1093/micmic/ozac043}
 }
 
 @article{li2021graph,
-  title={Graph network based deep learning of bandgaps},
+  title={Graph Network Based Deep Learning of Bandgaps},
   author={Li, Xiang-Guo and Blaiszik, Ben and Schwarting, Marcus Emory and Jacobs, Ryan and Scourtas, Aristana and Schmidt, KJ and Voyles, Paul M and Morgan, Dane},
   journal={The Journal of Chemical Physics},
   volume={155},
   number={15},
   pages={154702},
   year={2021},
-  publisher={AIP Publishing LLC}
+  publisher={AIP Publishing LLC},
+  doi={10.1063/5.0066009}
 }
 
 @article{schwalbe2021priori,
-  title={A priori control of zeolite phase competition and intergrowth with high-throughput simulations},
+  title={A Priori Control of Zeolite Phase Competition and Intergrowth with h\High-throughput Simulations},
   author={Schwalbe-Koda, Daniel and Kwon, Soonhyoung and Paris, Cecilia and Bello-Jurado, Estefania and Jensen, Zach and Olivetti, Elsa and Willhammar, Tom and Corma, Avelino and Rom{\'a}n-Leshkov, Yuriy and Moliner, Manuel and others},
   journal={Science},
   volume={374},
   number={6565},
   pages={308--315},
   year={2021},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.abh3350}
 }
 
 @article{dmc,
-  title={Towards DMC accuracy across chemical space with scalable $$\backslash$Delta $-QML},
+  title={Towards DMC Accuracy Across Chemical Space with Scalable $$\backslash$Delta $-QML},
   author={Huang, Bing and von Lilienfeld, O Anatole and Krogel, Jaron T and Benali, Anouar},
   journal={arXiv preprint arXiv:2210.06430},
-  year={2022}
+  year={2022},
[…]
