+++ JOSS/paper.bib
@@ -51,7 +51,7 @@
 }
 
 @article{schwalbe2021priori,
-  title={A Priori Control of Zeolite Phase Competition and Intergrowth with h\High-throughput Simulations},
+  title={A Priori Control of Zeolite Phase Competition and Intergrowth with High-throughput Simulations},
   author={Schwalbe-Koda, Daniel and Kwon, Soonhyoung and Paris, Cecilia and Bello-Jurado, Estefania and Jensen, Zach and Olivetti, Elsa and Willhammar, Tom and Corma, Avelino and Rom{\'a}n-Leshkov, Yuriy and Moliner, Manuel and others},
   journal={Science},
   volume={374},
