### 2024-01-12 b6db566
Update paper.bib with DOIs and correct citations (#422)
* Update paper.bib with DOIs and correct citations

* remove erroneous header notation
