### 2024-01-12 #5467 @ascourtas
@Fei-Tao Good evening!

1. I resolved all DOI and reference issues -- editorialbot has only one warning, but it's for an article that does not possess a DOI.
2. The most up-to-date code release can be found here: [v0.7.3](https://github.com/MLMI2-CSSI/foundry/releases/tag/v0.7.3)
3. The new archive of the code at the `joss` branch can be found here: [10.5281/zenodo.10494644](https://zenodo.org/doi/10.5281/zenodo.10494644)

Please let us know if any other changes need to be made. Thank you!
