+++ foundry/https_download.py
@@ -53,18 +53,37 @@ def _get_files(tc, ep, queue, max_depth):
 
 
 # TODO (wardlt): Avoid passing dictionaries, as documenting their content is tedious
-def download_file(item, https_config):
+def download_file(item, https_config, base_directory="data/"):
     """Download a file to disk
 
     Args:
         item: Dictionary defining the path to the file
         https_config: Configuration defining the URL of the server and the name of the dataset
     """
-    url = f"{https_config['base_url']}{item['path']}{item['name']}"
+    print('download_file')
+    base_url = https_config.get('base_url', '').rstrip('/')
+    path = item.get('path', '').strip('/')
+
+    # Extracting the name and subdirectory from the item
+    name = item.get('name', '')
+    subdirectory = name.split('/')[0] if '/' in name else ''
+
+    # Avoid duplication of subdirectory in path
+    if subdirectory and path.endswith(subdirectory):
+        full_path = f"{path}/{name.split('/', 1)[-1]}".strip('/')
+    else:
+        full_path = '/'.join([path, name]).strip('/')
+
+    url = f"{base_url}/{full_path}"
+
+    #url = f"{https_config['base_url']}{item['path']}{item['name']}"
+    print(item['path'])
+    print(item['name'])
+    print(f'Full Path:{full_path}')
+    print(f'URL: {url}')
 
     # build destination path for data file
-    destination = os.path.join("data/", https_config['source_id'], item['name'])
-
+    destination = os.path.join(base_directory, https_config['source_id'], item['name'])
     parent_path = os.path.split(destination)[0]
 
     # if parent directories don't exist, create them
@@ -79,3 +98,128 @@ def download_file(item, https_config):
 
     # TODO (wardlt): Should we just return the key?
     return {destination + " status": True}
+
+# def download_file(item, https_config, base_directory="data/", timeout=1800):
+#     """
+#     Download a file to disk with progress feedback.
+
+#     Args:
+#         item (dict): Dictionary defining the path to the file.
+#         https_config (dict): Configuration defining the URL of the server and the dataset name.
+#         base_directory (str): Base directory for downloads. Default is "data/".
+#         timeout (int): Timeout for the HTTPS request in seconds. Default is 30.
+
+#     Returns:
+#         str: The path to the downloaded file or None if the download failed.
+#     """
+#     print(https_config)
+
+#     # Validate https_config and item
+#     if 'base_url' not in https_config or 'source_id' not in https_config:
+#         print("Error: Missing required keys in https_config")
+#         return None
+#     if 'path' not in item or 'name' not in item:
+#         print("Error: Missing required keys in item")
+#         return None
+
+#     url = f"{https_config['base_url']}{item['path']}{item['name']}"
+#     destination = os.path.join(base_directory, https_config['source_id'], item['name'])
+#     parent_path = os.path.dirname(destination)
+
+#     print(f"URL: {url}  \n Destination: {destination} \n Parent Path: {parent_path}")
+
+#     os.makedirs(parent_path, exist_ok=True)
+
+#     try:
+#         with requests.get(url, stream=True, timeout=timeout) as response:
+#             response.raise_for_status()
+
+#             # Get the total file size from headers
+#             total_size = int(response.headers.get('content-length', 0))
+#             downloaded_size = 0
+
+#             with open(destination, "wb") as f:
+#                 for chunk in response.iter_content(chunk_size=8192):
+#                     if chunk:
+#                         f.write(chunk)
+#                         downloaded_size += len(chunk)
+#                         # Calculate and print the download progress
+#                         progress = (downloaded_size / total_size) * 100
+#                         print(f"\rDownloading... {progress:.2f}%", end="")
+#             print("\nDownload complete.")
[…]
