### 2024-01-14 6edb000
Minimizing requirements
