+++ .github/workflows/tests.yml
@@ -20,9 +20,9 @@ jobs:
       CLIENT_SECRET: ${{ secrets.CLIENT_SECRET }}
     name: build
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
       - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
           cache : 'pip'
