### 2024-01-14 61bf289
adding Pydantic typing for local_cache_dir
