### 2024-01-14 eea7dc7
Adding Pydantic typing for index
