### 2024-01-15 0ebdf89
DLHub Removal + making https download more robust (#423)
* Removing DLHub and fixing https download bug

* Add https streaming download with monitor + flake8 fixes

* make it easier to test locally with confidential client credentials

* Minimizing requirements

* updating GitHub Actions to latest versions (there were breaking issues with caching)

* Pinning pydantic less than 2
