+++ foundry/foundry.py
@@ -5,7 +5,7 @@ from json2table import convert
 import numpy as np
 import pandas as pd
 from pydantic import ValidationError
-from typing import Any, Dict, List
+from typing import Any, Dict, List, AnyStr
 import logging
 import warnings
 import os
@@ -14,7 +14,6 @@ from tqdm.auto import tqdm
 
 from mdf_connect_client import MDFConnectClient
 from mdf_forge import Forge
-from dlhub_sdk import DLHubClient
 from globus_sdk import AuthClient
 
 from .auth import PubAuths
@@ -37,15 +36,14 @@ class Foundry(FoundryBase):
     """Foundry Client Base Class
 
     Foundry object used for all interactions with Foundry datasets and models. Interfaces with MDF Connect Client,
-        Globus Compute, Globus Auth, Globus Transfer, Globus Search, DLHub, and relevant Globus Endpoints
+        Globus Compute, Globus Auth, Globus Transfer, Globus Search, and relevant Globus Endpoints
     """
 
-    dlhub_client: Any
     forge_client: Any
     connect_client: Any
     transfer_client: Any
     auth_client: Any
-    index = ""
+    index: AnyStr = ""
     auths: Any
 
     def __init__(
@@ -84,9 +82,7 @@ class Foundry(FoundryBase):
                     "search",
                     "petrel",
                     "transfer",
-                    "dlhub",
                     "openid",
-                    "https://auth.globus.org/scopes/facd7ccc-c5f4-42aa-916b-a0e270e2c2a9/all",  # funcx
                     "https://auth.globus.org/scopes/f10a69a9-338c-4e5b-baa1-0dc92359ab47/https",  # Eagle HTTPS
                     "https://auth.globus.org/scopes/82f1b5c6-6e9b-11e5-ba47-22000b92c6ec/https",  # NCSA HTTPS
                     "https://auth.globus.org/scopes/d31d4f5d-be37-4adc-a761-2f716b7af105/action_all",  # Globus Search Lambda
@@ -115,7 +111,7 @@ class Foundry(FoundryBase):
             search_client=self.auths["search"],
             transfer_client=self.auths["transfer"],
             data_mdf_authorizer=self.auths["data_mdf"],
-            petrel_authorizer=self.auths["petrel"],
+            petrel_authorizer=self.auths["petrel"]
         )
 
         self.transfer_client = self.auths['transfer']
@@ -130,19 +126,6 @@ class Foundry(FoundryBase):
 
         self.connect_client = MDFConnectClient(
             authorizer=self.auths["mdf_connect"], test=test
-        )
-
-        self.dlhub_client = DLHubClient(
-            dlh_authorizer=self.auths["dlhub"],
-            search_authorizer=self.auths["search_authorizer"],
-            fx_authorizer=self.auths[
-                "https://auth.globus.org/scopes/facd7ccc-c5f4-42aa-916b-a0e270e2c2a9/all"
-            ],
-            openid_authorizer=self.auths['openid'],
-            sl_authorizer=self.auths[
-                "https://auth.globus.org/scopes/d31d4f5d-be37-4adc-a761-2f716b7af105/action_all"
-            ],
-            force_login=False,
         )
 
     def load(self, name, download=True, globus=False, verbose=False, metadata=None, authorizers=None, **kwargs):
@@ -249,22 +232,6 @@ class Foundry(FoundryBase):
                 year, and DOI
         """
         return self.search()
-
-    def run(self, name, inputs, funcx_endpoint=None, **kwargs):
-        """Run a model on inputted data
-
-        Args:
-            name (str): DLHub model name
-            inputs: Data to send to DLHub as inputs (should be JSON serializable, example types include dict, list,
-                np.ndarray, etc)
[…]
