+++ requirements.txt
@@ -1,13 +1,10 @@
 globus-sdk>=3,<4
-dlhub_sdk>=2.0.3
 requests>=2.18.4
-tqdm>=4.19.4
 six>=1.11.0
 h5py>=2.10.0
 numpy>=1.15.4
 pandas>=0.23.4
-scikit-learn>=1.0
-pydantic>=1.6.1
+pydantic<2.0.0
 mdf_forge>=0.8.0
 mdf-connect-client>=0.4.0
 json2table>=1.1.5

[…]
