### 2024-01-15 a370a94
Update version in setup.py for release
