### 2024-01-18 77c79d9
tighten up testing, remove downloads from GHA, hold off on https download until next issue
