+++ tests/test_foundry.py
@@ -242,12 +242,12 @@ def test_search():
 
     assert isinstance(ds, pd.DataFrame)
     assert len(ds) > 0
-    
+
     dataset = ds.iloc[0].FoundryDataset
 
     # assert ds.iloc[0]['name'] is not None
     assert dataset.dc["titles"][0]["title"] is not None
-    
+
     # assert ds.iloc[0]['source_id'] is not None
     assert dataset.dataset_name is not None
 
@@ -268,12 +268,12 @@ def test_search_as_list():
 
     assert isinstance(ds, list)
     assert len(ds) > 0
-    
+
     dataset = ds[0]
 
     # assert ds[0]['name'] is not None
     assert dataset.dc["titles"][0]["title"] is not None
-    
+
     # assert ds[0]['source_id'] is not None
     assert dataset.dataset_name is not None
 
@@ -287,12 +287,12 @@ def test_search_limit():
 
     assert isinstance(ds, pd.DataFrame)
     assert len(ds) == 10
-    
+
     dataset = ds.iloc[0].FoundryDataset
 
     # assert ds.iloc[0]['name'] is not None
     assert dataset.dc["titles"][0]["title"] is not None
-    
+
     # assert ds.iloc[0]['source_id'] is not None
     assert dataset.dataset_name is not None
 
@@ -316,6 +316,7 @@ def test_metadata_pull():
     assert dataset.dc["titles"][0]["title"] == expected_title
 
 
+@pytest.mark.skipif(bool(is_gha), reason="Test does not succeed on GHA - no Globus endpoint")
 def test_download_globus():
     f = foundry.Foundry(globus=False, authorizers=auths)
     dataset = f.search(test_dataset).iloc[0].FoundryDataset
@@ -329,11 +330,17 @@ def test_download_globus():
     _delete_test_data(dataset)
 
 
-@pytest.mark.skip(reason='Not sure why, but http download is bringing up a globus auth page instead of a datafile')
+@pytest.mark.skip("Skipping until we can get HTTPS working on new PR")
 def test_download_https():
     f = foundry.Foundry(globus=False, authorizers=auths)
-    dataset = f.search(test_dataset)[0].FoundryDataset
-    assert dataset.dc["titles"][0]["title"] == expected_title
+    dataset = f.search(test_dataset).iloc[0].FoundryDataset
+    res = dataset.get_as_dict()
+    X, y = res['train']
+
+    assert len(X) > 1
+    assert isinstance(X, pd.DataFrame)
+    assert len(y) > 1
+    assert isinstance(y, pd.DataFrame)
     _delete_test_data(dataset)
 
 
@@ -344,7 +351,7 @@ def test_delete_cache():
     with mock.patch.object(builtins, 'input', lambda _: 'y'):
         dataset.clear_dataset_cache()
 
-    assert os.path.exists(os.path.join(dataset._foundry_cache.local_cache_dir, dataset.dataset_name)) == False
+    assert os.path.exists(os.path.join(dataset._foundry_cache.local_cache_dir, dataset.dataset_name)) is False
 
 
 @pytest.mark.skip(reason='Saving for #401')
@@ -366,7 +373,7 @@ def test_dataframe_load_split_wrong_split_name():
     f = foundry.Foundry(download=True, globus=False, authorizers=auths)
 
     with pytest.raises(Exception) as exc_info:
-        f.load_data(splits=['chewbacca'])
+        dataset = f.load_data(splits=['chewbacca'])
 
     err = exc_info.value
     assert hasattr(err, '__cause__')
@@ -376,14 +383,14 @@ def test_dataframe_load_split_wrong_split_name():
 
[…]
