+++ JOSS/paper.bib
@@ -30,6 +30,18 @@
   doi={10.1007/s11837-016-2001-3}
 }
 
+@article{wei2021benchmark,
+  title={Benchmark Tests of Atom-Locating CNN Models with a Consistent Dataset},
+  author={Wei, Jingrui and Blaiszik, Ben and Morgan, Dane and Voyles, Paul},
+  journal={Microscopy and Microanalysis},
+  volume={27},
+  number={S1},
+  pages={2518--2520},
+  year={2021},
+  publisher={Cambridge University Press},
+  doi={10.1017/S1431927621008989}
+}
+
 @article{wei2022benchmark,
   title={Benchmark Tests of Atom Segmentation Deep Learning Models with a Consistent Dataset},
   author={Wei, Jingrui and Blaiszik, Ben and Scourtas, Aristana and Morgan, Dane and Voyles, Paul M},
