+++ JOSS/paper.md
@@ -110,7 +110,7 @@ bibliography: paper.bib
 
 # Summary
 
-The application of open science and machine learning to scientific, engineering, and industry-relevant problems is a critical component of the cross-department U.S. Artificial Intelligence (AI) strategy highlighted e.g., by the AI Initiative, the recent National AI Strategy report [@nairr], the Year of Open Data [@yearofopendata], Materials Genome Initiative [@mgidepablo;@ward2015materials], and more. A key aspect of these strategies is to ensure that infrastructure exists to make datasets easily accessible for training, retraining, reproducing, and verifying model performance on chosen tasks. However, the discovery of high-quality, curated datasets adhering to the FAIR principles (findable, accessible, interoperable and reusable) remains a challenge.
+The application of open science and machine learning to scientific, engineering, and industry-relevant problems is a critical component of the cross-department U.S. Artificial Intelligence (AI) strategy highlighted e.g., by the AI Initiative, the recent National AI Strategy report [@nairr], the Year of Open Data, Materials Genome Initiative [@mgidepablo;@ward2015materials], and more. A key aspect of these strategies is to ensure that infrastructure exists to make datasets easily accessible for training, retraining, reproducing, and verifying model performance on chosen tasks. However, the discovery of high-quality, curated datasets adhering to the FAIR principles (findable, accessible, interoperable and reusable) remains a challenge.
 
 To overcome these dataset access challenges, we introduce Foundry-ML, software that combines several services to provide researchers capabilities to publish and discover structured datasets for ML in science, specifically in materials science and chemistry. Foundry-ML consists of a Python client, a web app, and standardized metadata and file structures built using services including the Materials Data Facility[@mdf-2019; @mdf-2016] and Globus [@ananthakrishnan2018globus; @chard2015globus]. Together, these services work in conjunction with Python software tooling to dramatically simplify data access patterns, as we show below.
 
