### 2024-01-19 f7dd124
remove erroneous spaces in fig ref
