+++ JOSS/paper.md
@@ -133,7 +133,7 @@ In addition to the Python software interface to each dataset, we have developed 
 
 # Usage
 Foundry has been successfully used in educational curricula [@foundrynanohub]
-and to publish datasets by research teams at the University of Chicago, Argonne National Lab, the University of Toronto [@dmc], 3M [@mmm], the University of Wisconsin [@wei2021benchmark; @li2021graph], MIT [@schwalbe2021priori] \autoref{fig 2}, and many more. In \autoref{fig 2}, we highlight a use case for the ML-guided design of organic structure–directing agents (OSDAs) to promote zeolite formation from the team of Gomez-Bombarelli at MIT. By using only the Foundry-ML software and the dataset DOI \autoref{fig1}a, which could be cited in a paper or retrieved from the Foundry-ML web app or software, a researcher can load descriptive metadata \autoref{fig1}b to understand the dataset contents, and load the data \autoref{fig1}c for analysis, exploration, and replication. A notebook showcasing this use case is available at in the GitHub examples linked in the Documentation section below.     
+and to publish datasets by research teams at the University of Chicago, Argonne National Lab, the University of Toronto [@dmc], 3M [@mmm], the University of Wisconsin [@wei2021benchmark; @li2021graph], MIT [@schwalbe2021priori] \autoref{fig2}, and many more. In \autoref{fig2}, we highlight a use case for the ML-guided design of organic structure–directing agents (OSDAs) to promote zeolite formation from the team of Gomez-Bombarelli at MIT. By using only the Foundry-ML software and the dataset DOI \autoref{fig1}a, which could be cited in a paper or retrieved from the Foundry-ML web app or software, a researcher can load descriptive metadata \autoref{fig1}b to understand the dataset contents, and load the data \autoref{fig1}c for analysis, exploration, and replication. A notebook showcasing this use case is available at in the GitHub examples linked in the Documentation section below.     
 
 # Future Directions
 In future work, we intend to add capabilities to Foundry-ML that enable publication and connection of datasets with ML models creating a combined ecosystem of datasets and models. This work will be completed in collaboration between two National Science Foundation (NSF) projects, (#1931306) “Collaborative Research: Framework: Machine Learning Materials Innovation Infrastructure” and (#2209892) "Garden: A FAIR Framework for Publishing and Applying AI Models for Translational Research in Science, Engineering, Education, and Industry". We also plan to generalize the metadata and extend these capabilities to scientific datasets in domains beyond materials and chemistry.
