### 2024-01-19 #5467 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
