### 2024-01-19 MLMI2-CSSI/foundry#419 @ascourtas
Additional steps were required:

- [ ] fix DOI issues
- [ ] reupload zip -- of `joss` branch this time, so need to redo it
- [ ] re-release (figure out what this means, since the `joss` branch is not a released branch)
