### 2024-01-19 MLMI2-CSSI/foundry#419 @ascourtas
More steps:

- [ ] fix citation errors
- [ ] reupload zip of `joss` branch to Zenodo
- [ ] get confirmation from JOSS
