### 2024-01-21 #5467 @Kevin-Mattheus-Moerman
@Fei-Tao thanks for editing here. It looks like you called `@editoralbot recommed-accept` a bit prematurely. Please first call `@editorialbot create post-review checklist` and ensure all final steps are completed first. In addition:

- [ ] The version is not set here.
- [ ] Above, I see comments asking if the joss or main branch  should  be archived. The archived version should be for the tagged release and need not include paper files. It is about the software reviewed and the subsequent tagged-release for this. Therefore please check that the archived version, they may need to create a new one, reflects the correct/latest reviewed software.
- [ ] Once the tagged release version is created and archived you should check that the same version tag is used on ZENODO (it currently isn't).
- [ ] The authors should add state, city, and country, to their affiliations. In addition, they should spell out USA as United States of America.
