### 2024-01-22 #5467 @ascourtas
Hello @Fei-Tao ! Per your request, I have completed the following:

- updated the Zenodo archive to [v0.7.4](https://github.com/MLMI2-CSSI/foundry/releases/tag/v0.7.4) with the latest code from `main`; it can still be found at the DOI [10.5281/zenodo.10494644](https://zenodo.org/doi/10.5281/zenodo.10494644) (resolves to latest)
- updated all author affiliations to include city, state, and unabbreviated country name
- confirmed all author names and Orcid IDs
- confirmed the correct license is set in Zenodo
- confirmed the correct title is used in Zenodo
