### 2024-01-22 #5467 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
