+++ tests/test_foundry.py
@@ -314,7 +314,7 @@ def test_search_no_results():
     assert hasattr(err, '__cause__')
 
 
-@pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #4')
+# @pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #4')
 def test_metadata_pull():
     f = foundry.Foundry(download=False, authorizers=auths)
     dataset = f.search(test_dataset).iloc[0].FoundryDataset
@@ -349,7 +349,7 @@ def test_download_https():
     _delete_test_data(dataset)
 
 
-@pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #2')
+# @pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #2')
 def test_delete_cache():
     f = foundry.Foundry(globus=True, authorizers=auths)
     dataset = f.search(test_dataset).iloc[0].FoundryDataset
@@ -396,7 +396,7 @@ def test_dataframe_load_split_but_no_splits():
     _delete_test_data(dataset)
 
 
-@pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #1')
+# @pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #1')
 def test_dataframe_search_by_doi():
     f = foundry.Foundry(globus=False, authorizers=auths)
 
