+++ tests/test_foundry.py
@@ -255,14 +255,12 @@ def test_search():
     assert dataset.dc.get("publicationYear", None) is not None
 
 
-# @pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #7')
 def test_dataset_get_citation():
     f = foundry.Foundry(authorizers=auths)
     ds = f.search(test_dataset).iloc[0].FoundryDataset
     assert ds.get_citation() is not None
 
 
-# @pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #6')
 def test_search_as_list():
     f = foundry.Foundry(authorizers=auths)
     q = "Elwood"
@@ -283,7 +281,7 @@ def test_search_as_list():
     assert dataset.dc.get("publicationYear", None) is not None
 
 
-@pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #8')
+# @pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #8')
 def test_search_limit():
     f = foundry.Foundry(authorizers=auths)
     ds = f.search(limit=10)
@@ -303,18 +301,6 @@ def test_search_limit():
     assert dataset.dc.get("publicationYear", None) is not None
 
 
-@pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #5')
-def test_search_no_results():
-    f = foundry.Foundry()
-
-    with pytest.raises(Exception) as exc_info:
-        f.search('chewbacca')
-
-    err = exc_info.value
-    assert hasattr(err, '__cause__')
-
-
-# @pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #4')
 def test_metadata_pull():
     f = foundry.Foundry(download=False, authorizers=auths)
     dataset = f.search(test_dataset).iloc[0].FoundryDataset
@@ -396,7 +382,6 @@ def test_dataframe_load_split_but_no_splits():
     _delete_test_data(dataset)
 
 
-# @pytest.mark.skipif(bool(is_gha), reason='Checking if tests work #1')
 def test_dataframe_search_by_doi():
     f = foundry.Foundry(globus=False, authorizers=auths)
 
